% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_utils.R
\name{Add2History}
\alias{Add2History}
\title{Add row to history}
\usage{
Add2History(history, step, substep, param.name, value)
}
\arguments{
\item{history}{A \code{data.frame} corresponding to the current history.}

\item{step}{A \code{character(1)} corresponding to the step name.}

\item{substep}{A \code{character(1)} corresponding to the substep name.}

\item{param.name}{A \code{character(1)} corresponding to the parameter name.}

\item{value}{The value of the corresponding parameter.}
}
\value{
A \code{data.frame} with one added row
}
\description{
This function adds a row to the history.
}
\examples{
history <- InitializeHistory()
Add2History(history, "Example step", "First sub-step", "my param", "THE value")

}
