% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_aggregate.R
\name{Add_Aggregated_rowData}
\alias{Add_Aggregated_rowData}
\title{Add_Aggregated_rowData()}
\usage{
Add_Aggregated_rowData(obj)
}
\arguments{
\item{obj}{An instance of \code{QFeatures} class}
}
\value{
An instance of \code{QFeatures} class
}
\description{
Add_Aggregated_rowData()
}
\author{
Samuel Wieczorek
}
