% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Add_Item_to_Dataset}
\alias{Add_Item_to_Dataset}
\title{Adds an instance of \code{SummarizedExperiment} to a \code{QFeatures} object}
\usage{
Add_Item_to_Dataset(dataset, name)
}
\arguments{
\item{dataset}{An instance of \code{SummarizedExperiment} class}

\item{name}{A \code{character()} for the new assay}
}
\value{
The dataset minus some items
}
\description{
Adds one or more items to the dataset. This function is specific of the
type of dataset.
}
