% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggregateIter}
\alias{aggregateIter}
\title{xxxx}
\usage{
aggregateIter(
  obj.pep,
  X,
  init.method = "Sum",
  method = "Mean",
  n = NULL,
  level
)
}
\arguments{
\item{obj.pep}{xxxxx}

\item{X}{xxxx}

\item{init.method}{xxxxx}

\item{method}{xxxxx}

\item{n}{xxxx}
}
\value{
A protein object of class \code{MSnSet}
}
\description{
xxxx
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
protID <- "Protein_group_IDs"
obj <- Exp1_R25_pept[[2]]
obj <- obj[seq_len(10)]
X <- BuildAdjacencyMatrix(obj)
ll.agg <- aggregateIter(obj.pep = obj)

}
\author{
Samuel Wieczorek
}
