% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggregateIterParallel}
\alias{aggregateIterParallel}
\title{xxxx}
\usage{
aggregateIterParallel(
  obj.pep,
  X,
  init.method = "Sum",
  method = "Mean",
  n = NULL,
  conds
)
}
\arguments{
\item{obj.pep}{xxxxx}

\item{X}{xxxx}

\item{init.method}{xxxxx}

\item{method}{xxxxx}

\item{n}{xxxx}
}
\value{
xxxxx
}
\description{
xxxx
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj.pep <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj.pep[[length(obj.pep)]])
conds <- colData(obj.pep)$Condition
obj.agg <- aggregateIterParallel(obj.pep = obj.pep[[length(obj.pep)]], X = X, n = 3, conds = conds)
}

}
\author{
Samuel Wieczorek
}
