% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggregateProstar2}
\alias{aggregateProstar2}
\title{Compute the intensity of proteins with the sum of the intensities
of their peptides.}
\usage{
aggregateProstar2(obj, i, FUN = "Sum", X)
}
\arguments{
\item{X}{An adjacency matrix in which lines and columns correspond
respectively to peptides and proteins.}

\item{obj.pep}{A matrix of intensities of peptides}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins based on the sum of the
intensities of their peptides.
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj.pep <- Exp1_R25_pept[seq_len(20)]
last.obj <- obj.pep[[length(obj.pep)]]
#obj.pep.imp <- wrapper.impute.detQuant(obj.pep[[length(obj.pep)]], na.type = c("Missing POV", "Missing MEC"))
X <- BuildAdjacencyMatrix(last.obj)
ll.agg <- aggregateProstar2(obj.pep, length(obj.pep), X = X)
}

}
\author{
Alexia Dorffer
}
