% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggregateTopn}
\alias{aggregateTopn}
\title{Compute the intensity of proteins as the sum of the
intensities of their n best peptides.}
\usage{
aggregateTopn(obj.pep, i, method = "Mean", n = 10, X)
}
\arguments{
\item{obj.pep}{A matrix of intensities of peptides}

\item{method}{xxx}

\item{n}{The maximum number of peptides used to aggregate a protein.}

\item{X}{An adjacency matrix in which lines and columns correspond
respectively to peptides and proteins.}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins as the sum of the
intensities of their n best peptides.
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
obj.pep <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj.pep[[length(obj.pep)]])
adjacencyMatrix(obj.pep[[length(obj.pep)]]) <- X
conds <- colData(obj.pep)$Condition
ll.agg <- aggregateTopn(obj.pep[[length(obj.pep)]], n = 3, conds = conds)

}
\author{
Alexia Dorffer, Samuel Wieczorek
}
