% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{applyAnovasOnProteins}
\alias{applyAnovasOnProteins}
\title{iteratively applies OWAnova() on the features of an MSnSet object}
\usage{
applyAnovasOnProteins(obj, i)
}
\arguments{
\item{obj}{a QFreatures object}

\item{i}{An integer which is the index of the assay in the QFeatures object
'}
}
\value{
a list of linear models
}
\description{
iteratively applies OWAnova() on the features of an MSnSet object
}
\examples{
data(subR25prot)
applyAnovasOnProteins(subR25prot[1:5,], 1)

}
\author{
Thomas Burger
}
