% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_aggregate.R
\name{BuildColumnToProteinDataset}
\alias{BuildColumnToProteinDataset}
\title{creates a column for the protein dataset after agregation by
using the previous peptide dataset.}
\usage{
BuildColumnToProteinDataset(peptideData, matAdj, columnName, proteinNames)
}
\arguments{
\item{peptideData}{A data.frame of meta data of peptides. It is the rowData
of the SummarizedExperiment object.}

\item{matAdj}{The adjacency matrix used to agregate the peptides data.}

\item{columnName}{The name(s) of the column in Biobase::rowData(peptides_MSnset)
that the user wants to keep in the new protein data.frame.}

\item{proteinNames}{The names of the protein in the new dataset
(i.e. rownames)}
}
\value{
A vector
}
\description{
This function creates a column for the protein dataset after aggregation
by using the previous peptide dataset.
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[1:10]
protID <- parentProtId(obj[[2]])
X <- adjacencyMatrix(obj[[2]])

X.split <- DaparToolshed::splitAdjacencyMat(X)
X.shared <- X.split$Xshared
X.unique <- X.split$Xspec


#adjacencyMatrix(obj[[2]]) <- X.unique
#rowdata.pep <- rowData(obj[[2]])


# obj <- aggregateFeatures4Prostar(
#   object = obj,
#   i = length(obj),
#   name = 'aggregated',
#   fcol = 'adjacencyMatrix',
#   fun = 'colSumsMat')
# 
# 
# .names <- "Sequence"
# 
# proteinNames <- rownames(obj[[length(obj)]])
# data <- rowData(obj[[length(obj)-1]])
# 
# new.col <- BuildColumnToProteinDataset(
#   peptideData = rowData(obj[[length(obj)-1]]), 
#   matAdj = adjacencyMatrix(obj[[2]]), 
#   columnName = "Sequence",
#   proteinNames = rownames(obj[[length(obj)]]))
}
\author{
Samuel Wieczorek
}
