% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{buildGraph}
\alias{buildGraph}
\title{Display a CC}
\usage{
buildGraph(The.CC, X)
}
\arguments{
\item{The.CC}{A cc (a list)}

\item{X}{xxxxx}
}
\value{
A plot
}
\description{
Display a CC
}
\examples{
data(subR25pept)
X <- QFeatures::adjacencyMatrix(subR25pept[[1]])
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)

}
\author{
Thomas Burger, Samuel Wieczorek
}
