% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{BuildMetacell}
\alias{BuildMetacell}
\title{Metacell function}
\usage{
BuildMetacell(from = NULL, level, qdata = NULL, conds = NULL, df = NULL)
}
\arguments{
\item{from}{A string designing the software used, either "maxquant", "proline" or "DIA-NN"}

\item{level}{A string designing the type of entity/pipeline.
Available values are: \code{peptide}, \code{protein}}

\item{qdata}{A matrix of quantitative data}

\item{conds}{A 1-col dataframe with the condition associated to each sample.}

\item{df}{A data.frame which contains the type of identification of the
entities. It must have the same dimensions as \code{qData}.}
}
\value{
NA
}
\description{
Create metacell
}
\examples{
data(subR25pept)
conds <- design.qf(subR25pept)$Condition
qdata <- SummarizedExperiment::assay(subR25pept[[2]])
metacell1 <- BuildMetacell('maxquant', 'peptide', qdata, conds)
metacell2 <- BuildMetacell('proline', 'peptide', qdata, conds)
}
\author{
Samuel Wieczorek
}
