% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{checkDesign}
\alias{checkDesign}
\title{Check if the design is valid}
\usage{
checkDesign(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the \code{pData()} function
of package \code{MSnbase}.}
}
\value{
A boolean
}
\description{
Check if the design is valid
}
\examples{
data(subR25pept)
checkDesign(SummarizedExperiment::colData(subR25pept)[, -1])

}
\author{
Thomas Burger, Samuel Wieczorek
}
