% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CleanRowData}
\alias{CleanRowData}
\title{Clean row data}
\usage{
CleanRowData(obj, i)
}
\arguments{
\item{obj}{An instance of the class \code{QFeatures}}

\item{i}{An integer which is the index of the assay in the QFeatures object}
}
\value{
An instance of the SummarizedExperiment structure
}
\description{
Clean row data
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
