% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.t.tests.R
\name{compute_t_tests}
\alias{compute_t_tests}
\title{Compute a t-test}
\usage{
compute_t_tests(
  obj,
  i = 1,
  contrast = c("OnevsOne", "OnevsAll"),
  type = c("Student", "Welch")
)
}
\arguments{
\item{obj}{A matrix of quantitative data, without any missing values.}

\item{i}{An integer which is the index of the assay in the QFeatures object}

\item{contrast}{Indicates if the test consists of the comparison of each
biological condition versus
each of the other ones (contrast=1;
for example H0:"C1=C2" vs H1:"C1!=C2", etc.)
or each condition versus all others (contrast=2; e.g.  H0:"C1=(C2+C3)/2" vs
H1:"C1!=(C2+C3)/2", etc. if there are three conditions).}

\item{type}{Either "Student" or "Welch"}
}
\value{
A list of two items : logFC and P_Value; both are dataframe. The
first one contains the logFC values of all the comparisons (one column for
one comparison), the second one contains the pvalue of all the comparisons
(one column for one comparison). The names of the columns for those two
dataframes are identical and correspond to the description of the comparison.
}
\description{
Compute a t-test
}
\examples{
library(SummarizedExperiment)
data(subR25prot)
obj <- subR25prot
filter <- FunctionFilter('qMetacellOnConditions',
  cmd = 'delete',
  mode = 'AtLeastOneCond',
  pattern = c("Missing POV", "Missing MEC"),
  conds = design_qf(obj)$Condition,
  percent = TRUE,
  th = 0.65,
  operator = '>')
obj <- filterFeaturesOneSE(obj, name = "Filtered", filters = list(filter))
ttest <- compute_t_tests(obj, 2)

}
\author{
Florence Combes, Samuel Wieczorek
}
