% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggQmetacell}
\alias{aggQmetacell}
\alias{aggregateMethods}
\alias{BuildColumnToProteinDataset}
\alias{Add_Aggregated_rowData}
\alias{metacell_agg}
\alias{select_topn}
\alias{innerAggregateIter}
\alias{innerSum}
\alias{innerMean}
\alias{innerMedian}
\alias{innerMedianpolish}
\alias{innerRobustsummary}
\title{Aggregate metacell}
\usage{
aggQmetacell(qMeta, X, level, conds)

aggregateMethods()

BuildColumnToProteinDataset(peptideData, matAdj, columnName, proteinNames)

Add_Aggregated_rowData(obj, col, i.agg)

metacell_agg(aggregatedSE, originalSE, adj_mat, conds, protname_order)

select_topn(pepData, X, n = 10, funpept = "Mean")

innerAggregateIter(
  pepData,
  X,
  init.method = "Mean",
  method = "Mean",
  n = NULL,
  uniqueiter = FALSE,
  topn_fun = "Mean",
  max_iter = 500
)

innerSum(pepData, X)

innerMean(pepData, X)

innerMedian(pepData, X)

innerMedianpolish(pepData, X)

innerRobustsummary(pepData, X)
}
\arguments{
\item{qMeta}{A \code{matrix} with quantitative metadata tag.}

\item{X}{A \code{matrix} acting as an adjacency matrix.}

\item{level}{A \code{character(1)} which is the type of dataset}

\item{conds}{A \code{character()} vector which is the names of conditions
for each sample in the dataset.}

\item{peptideData}{A data.frame of meta data of peptides. It is the rowData
of the SummarizedExperiment object.}

\item{matAdj}{The adjacency matrix used to aggregate the peptides data.}

\item{columnName}{The name(s) of the column in SummarizedExperiment::rowData(peptides_MSnset)
that the user wants to keep in the new protein data.frame.}

\item{proteinNames}{The names of the protein in the new dataset
(i.e. rownames)}

\item{obj}{An instance of class \link[QFeatures:QFeatures]{QFeatures}.}

\item{col}{A \code{character()} of column names from rowdata to be aggregated.}

\item{i.agg}{A \code{numeric(1)} indicating the index of the assay to which add the aggregated rowData, using the previous assay's rowData.}

\item{aggregatedSE}{An instance of class \link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment} containing the aggregated data.}

\item{originalSE}{An instance of class \link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment} containing the non-aggregated data.}

\item{adj_mat}{An adjacency matrix.}

\item{protname_order}{A \code{character()} vector with the protein name in order.}

\item{pepData}{A \code{matrix} containing the peptide intensities.
Note that the function assume that data is already log-transformed.}

\item{n}{A \code{numeric(1)} specifying the number of peptides to use for each protein. If \code{NULL}, all peptides are considered.}

\item{funpept}{A function used for determining a peptide's value.
Available functions are \code{Sum}, \code{Mean} or \code{Median}.}

\item{init.method}{A function used for initializing the aggregation.
Available functions are \code{Sum}, \code{Mean}, \code{Median}, \code{medianPolish} or \code{robustSummary}.
See below for details.}

\item{method}{A function used for the aggregation.
Available functions are \code{Sum}, \code{Mean}, \code{Median}, \code{medianPolish} or \code{robustSummary}.
See below for details.}

\item{uniqueiter}{A bole}

\item{topn_fun}{A function used to determine how to choose the top n peptides.
Available functions are \code{Sum}, \code{Mean} or \code{Median}.
See below for details.}

\item{max_iter}{A \code{numeric(1)} setting the maximum number of iteration.}
}
\value{
NA

A vector

An instance of \code{QFeatures} class with aggregated rowData in specified assay.

A \code{SummarizedExperiment} containing the aggregated data.

An adjacency matrix with only the top n peptides selected.

A \code{matrix} containing the aggregated values.

A \code{matrix} containing the aggregated values.

A \code{matrix} containing the aggregated values.

A \code{matrix} containing the aggregated values.

A \code{matrix} containing the aggregated values.

A \code{matrix} containing the aggregated values.
}
\description{
Aggregate the quantitative metadata tag.

This function creates a column for the protein dataset after aggregation
by using the previous peptide dataset.

Aggregation of rowData of a \code{QFeatures} assay.

Aggregate the metadata

This function aggregate quantitative data using a method of redistribution of shared peptides. Intensity of shared peptides are redistributed proportionally to each protein.
Note that the function assumes that the intensities are not log-transformed.

Aggregation using sum method.

Aggregation using mean method.

Aggregation using median method.

Aggregation using medianPolish method.
Note that this method is parallelized to be more efficient.

Aggregation using robustSummary method.
}
\details{
Available functions are :
\itemize{
\item \code{Sum} : base::rowSums()
\item \code{Mean} : base::rowMeans()
\item \code{Median} : matrixStats::rowMedians()
\item \code{medianPolish} : MsCoreUtils::medianPolish(), not available for \code{topn_fun}.
Note that this method takes significantly more time than the others, and is parallelized to be more efficient.
\item \code{robustSummary} : MsCoreUtils::robustSummary(), not available for \code{topn_fun}.
Note that this method takes significantly more time than the others, and is parallelized to be more efficient.
}
}
\examples{
data(subR25pept)
qMeta <- qMetacell(subR25pept, 1)
X <- QFeatures::adjacencyMatrix(subR25pept[[1]])
level <- typeDataset(subR25pept[[1]])
conds <- SummarizedExperiment::colData(subR25pept)$Condition
aggQmeta <- aggQmetacell(qMeta, X, level, conds)

library(QFeatures)

data(subR25pept)
protID <- parentProtId(subR25pept[[1]])
X <- QFeatures::adjacencyMatrix(subR25pept[[1]])

X.split <- DaparToolshed::splitAdjacencyMat(X)
X.shared <- X.split$Xshared
X.unique <- X.split$Xspec
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
X.topn <- select_topn(SummarizedExperiment::assay(subR25pept[[1]]), X, n = 3)

data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
qdata.agg <- innerAggregateIter(SummarizedExperiment::assay(subR25pept[[1]]), X)

library(QFeatures)
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
i.sum <- innerSum(SummarizedExperiment::assay(subR25pept[[1]]), X)

\donttest{
library(QFeatures)
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
i.mean <- innerMean(SummarizedExperiment::assay(subR25pept), X)
}

\donttest{
library(QFeatures)
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
i.mean <- innerMedian(SummarizedExperiment::assay(subR25pept[[1]]), X)
}

\donttest{
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
i.mean <- innerMedianpolish(SummarizedExperiment::assay(subR25pept[[1]]), X)
}

\donttest{
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
i.mean <- innerRobustsummary(SummarizedExperiment::assay(subR25pept[[1]]), X)
}

}
\author{
Samuel Wieczorek

Samuel Wieczorek, Manon Gaudin

Manon Gaudin
}
