% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaComputeAdjustedPValues}
\alias{diffAnaComputeAdjustedPValues}
\title{Computes the adjusted p-values}
\usage{
diffAnaComputeAdjustedPValues(pval, pi0Method = 1)
}
\arguments{
\item{pval}{The result (p-values) of the differential analysis}

\item{pi0Method}{The parameter pi0.method of the method adjust.p in the
package \code{cp4p}}
}
\value{
The computed adjusted p-values
}
\description{
This function is a wrapper to the function adjust.p from the \code{cp4p} package.
It returns the adjusted p-values corresponding to the p-values of the differential
analysis. The adjusted p-values is computed with the function \code{p.adjust}\{stats\}.
}
\examples{
data(subR25prot)
obj <- subR25prot
# Simulate imputation
obj <- NAIsZero(obj, 1)
obj <- NAIsZero(obj, 2)
allComp <- limmaCompleteTest(SummarizedExperiment::assay(obj[[length(obj)]]), design.qf(obj), comp.type="OnevsOne")
diffAnaComputeAdjustedPValues(pval = allComp$P_Value[, 1])

}
\author{
Samuel Wieczorek
}
