% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaComputeFDR}
\alias{diffAnaComputeFDR}
\title{Computes the FDR corresponding to the p-values of the
differential analysis}
\usage{
diffAnaComputeFDR(adj.pvals)
}
\arguments{
\item{adj.pvals}{The adjusted p-values of the differential analysis}
}
\value{
The computed FDR value (floating number)
}
\description{
This function returns the FDR corresponding to the p-values of the differential
analysis.
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
