% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_volcano.R
\name{diffAnaVolcanoplot_rCharts}
\alias{diffAnaVolcanoplot_rCharts}
\title{Volcanoplot of the differential analysis}
\usage{
diffAnaVolcanoplot_rCharts(
  df,
  th_pval = 1e-60,
  th_logfc = 0,
  conditions = NULL,
  clickFunction = NULL,
  pal = NULL
)
}
\arguments{
\item{df}{A dataframe which contains the following slots :
x : a vector of the log(fold change) values of the differential analysis,
y : a vector of the p-value values returned by the differential analysis.
index : a vector of the rowanmes of the data.
This dataframe must has been built with the option stringsAsFactors set
to FALSE. There may be additional slots which will be used to show
informations in the tooltip. The name of these slots must begin with the
prefix "tooltip_". It will be automatically removed in the plot.}

\item{th_pval}{A floating number which represents the p-value that
separates differential and non-differential data.}

\item{th_logfc}{A floating number which represents the log of the
Fold Change that separates differential and non-differential data.}

\item{conditions}{A list of the names of condition 1 and 2 used for the
differential analysis.}

\item{clickFunction}{A string that contains a JavaScript function used to
show info from slots in df. The variable this.index refers to the slot
named index and allows to retrieve the right row to show in the tooltip.}

\item{pal}{A \code{list} containing 2 color to use for the plot}
}
\value{
An interactive volcanoplot
}
\description{
#' Plots an interactive volcanoplot after the differential analysis.
Typically, the log of Fold Change is represented on the X-axis and the
log10 of the p-value is drawn on the Y-axis. When the \code{th_pval}
and the \code{th_logfc} are set, two lines are drawn respectively on
the y-axis and the X-axis to visually distinguish between differential and
non differential data. With the use of the package Highcharter, a
customizable tooltip appears when the user put the mouse's pointer over
a point of the scatter plot.
}
\examples{
\donttest{
data(subR25prot)
obj <- subR25prot
# Simulate imputation
obj <- NAIsZero(obj, 1)
allComp <- limmaCompleteTest(
SummarizedExperiment::assay(obj[[length(obj)]]), 
design.qf(obj), 
comp.type="OnevsOne")
df <- data.frame(
x = allComp$logFC[[1]],
y = -log10(allComp$P_Value[[1]]),
index = as.character(rownames(obj[[2]]))
)
tooltipSlot <- c("Fasta_headers", "Sequence_length")
df <- cbind(df, SummarizedExperiment::rowData(obj[[2]])[, tooltipSlot])
colnames(df) <- gsub(".", "_", colnames(df), fixed = TRUE)
if (ncol(df) > 3) {
  colnames(df)[4:ncol(df)] <- paste0("tooltip_", 
                                     colnames(df)[4:ncol(df)])
}
cond <- c("25fmol", "10fmol")
diffAnaVolcanoplot_rCharts(
  df,
  th_pval = 2.5,
  th_logfc = 1,
  conditions = cond
)
}

}
\author{
Samuel Wieczorek
}
