% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{displayCCvisNet}
\alias{displayCCvisNet}
\title{Display a CC}
\usage{
displayCCvisNet(g)
}
\arguments{
\item{g}{A cc (a list)}
}
\value{
A plot
}
\description{
Display a CC
}
\examples{
data(subR25pept)
X <- QFeatures::adjacencyMatrix(subR25pept[[1]])
ll <- getPepProtCC(X)
g <- buildGraph(ll[[1]], X)
displayCCvisNet(g)

}
\author{
Thomas Burger, Samuel Wieczorek
}
