% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{finalizeAggregation}
\alias{finalizeAggregation}
\title{Finalizes the aggregation process}
\usage{
finalizeAggregation(obj.pep, i, protData, protMetacell, X)
}
\arguments{
\item{obj.pep}{A peptide object of class \code{MSnSet}}

\item{protData}{xxxxx}

\item{protMetacell}{xxx}

\item{X}{An adjacency matrix in which lines and columns correspond
respectively to peptides and proteins.}

\item{pepData}{xxxx}
}
\value{
A protein object of class \code{MSnSet}
}
\description{
Method to finalize the aggregation process
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj.pep <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj.pep[[length(obj.pep)]])
adjacencyMatrix(obj.pep[[length(obj.pep)]]) <- X
}

}
\author{
Samuel Wieczorek
}
