% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{formatPHTResults}
\alias{formatPHTResults}
\title{Extracts the FC and the raw p-values and format them in a data.frame}
\usage{
formatPHTResults(post_hoc_models_summaries)
}
\arguments{
\item{post_hoc_models_summaries}{resulting from applying lapply(summary((.)) to a multicomp function.}
}
\value{
A \code{data.frame()}
}
\description{
Extracts the FC and the raw p-values and format them in a data.frame
}
\examples{
NULL


}
\author{
Thomas Burger
}
