% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{getDesignLevel}
\alias{getDesignLevel}
\title{xxx}
\usage{
getDesignLevel(sTab)
}
\arguments{
\item{sTab}{xxx}
}
\description{
xxx
}
\examples{
data(subR25pept)
sTab <- SummarizedExperiment::colData(subR25pept)
getDesignLevel(sTab)

}
