% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{GetDetailedNbPeptidesUsed}
\alias{GetDetailedNbPeptidesUsed}
\title{Computes the detailed number of peptides used for aggregating
each protein}
\usage{
GetDetailedNbPeptidesUsed(pepData, X)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A list of two items
}
\description{
Method to compute the detailed number of quantified peptides used for
aggregating each protein
}
\examples{
library(SummarizedExperiment)
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
ll.n <- GetDetailedNbPeptidesUsed(SummarizedExperiment::assay(subR25pept), X)

}
\author{
Samuel Wieczorek
}
