% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetHistory}
\alias{GetHistory}
\title{Wrapper for the \code{paramshistory} function.}
\usage{
GetHistory(obj.se, history)
}
\arguments{
\item{obj.se}{An instance of the class \code{SummarizedExperiment}}

\item{history}{A \code{data.frame()}}
}
\value{
A \code{data.frame()}
}
\description{
Wrapper for the \code{paramshistory} function.
}
\examples{
data(subR25prot)
GetHistory(subR25prot[[1]])

}
\author{
Samuel Wieczorek
}
