% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_FunctionFiltering.R
\name{GetIndices_WholeLine}
\alias{GetIndices_WholeLine}
\title{Search lines which respects query on all their elements.}
\usage{
GetIndices_WholeLine(metacell.mask)
}
\arguments{
\item{metacell.mask}{A \code{data.frame()} of boolean values which indictes the
presence (TRUE) or not (FALSE) of given tags in the quantitative cells
metadata}
}
\value{
A vector of \code{integer()}
}
\description{
This function looks for the lines where each element respect the query.
}
\examples{
data(subR25pept)
level <- 'peptide'
pattern <- "Missing POV"
metacell.mask <- match.metacell(metadata = qMetacell(subR25pept[[1]]), 
pattern = pattern, level = level)
ind <- GetIndices_WholeLine(metacell.mask)

}
