% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{GetNbPeptidesUsed}
\alias{GetNbPeptidesUsed}
\title{Computes the number of peptides used for aggregating each protein}
\usage{
GetNbPeptidesUsed(pepData, X)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A data.frame
}
\description{
Method to compute the number of quantified peptides used for aggregating
each protein
}
\examples{
\donttest{
library(QFeatures)
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
GetNbPeptidesUsed(SummarizedExperiment::assay(subR25pept), X)
}

}
\author{
Samuel Wieczorek
}
