% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{GetNbTags}
\alias{GetNbTags}
\title{Number of each metacell tags}
\usage{
GetNbTags(obj)
}
\arguments{
\item{obj}{A instance of the class \code{SummarizedExperiment}}
}
\value{
An integer
}
\description{
Number of each metacell tags
}
\examples{
library(DaparToolshed)
data(subR25prot, package = 'DaparToolshed')
GetNbTags(subR25prot[[1]])

}
