% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_missingValuesFiltering.R
\name{getNumberOf}
\alias{getNumberOf}
\title{Number of lines with prefix}
\usage{
getNumberOf(obj, name = NULL, prefix = NULL)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment}.}

\item{name}{The name of a column.}

\item{prefix}{A string}
}
\value{
An integer
}
\description{
Returns the number of lines, in a given column, where content matches
the prefix.
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
getNumberOf(Exp1_R25_pept[[1]], "Potential_contaminant", "+")

}
\author{
Samuel Wieczorek
}
