% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{getPepProtCC}
\alias{getPepProtCC}
\title{Build the list of connex composant of the adjacency matrix}
\usage{
getPepProtCC(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of CC
}
\description{
Build the list of connex composant of the adjacency matrix
}
\examples{
data(subR25pept)
X <- QFeatures::adjacencyMatrix(subR25pept[[1]])
ll <- getPepProtCC(X)

}
\author{
Thomas Burger, Samuel Wieczorek
}
