% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_missingValuesFiltering.R
\name{getPourcentageOfMV}
\alias{getPourcentageOfMV}
\title{Percentage of missing values}
\usage{
getPourcentageOfMV(obj)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment}.}
}
\value{
A floating number
}
\description{
Returns the percentage of missing values in the quantitative data.
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
getPourcentageOfMV(Exp1_R25_pept[[1]])

}
\author{
Florence Combes, Samuel Wieczorek
}
