% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{getProteinsStats}
\alias{getProteinsStats}
\title{Computes the number of proteins that are only defined by
specific peptides, shared peptides or a mixture of two.}
\usage{
getProteinsStats(X)
}
\arguments{
\item{X}{The adjacency matrix with both specific and
shared peptides.}
}
\value{
A list
}
\description{
This function computes the number of proteins that are only defined by
specific peptides, shared peptides or a mixture of two.
}
\examples{
data(subR25pept)
obj.last <- subR25pept[[1]]
X <- BuildAdjacencyMatrix(subR25pept[[1]])
getProteinsStats(X)


}
\author{
Samuel Wieczorek, Manon Gaudin
}
