% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{GraphPepProt}
\alias{GraphPepProt}
\title{Function to create a histogram that shows the repartition of
peptides w.r.t. the proteins}
\usage{
GraphPepProt(mat)
}
\arguments{
\item{mat}{An adjacency matrix.}
}
\value{
A histogram
}
\description{
Method to create a plot with proteins and peptides on
a MSnSet object (peptides)
}
\examples{
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
GraphPepProt(X)

}
\author{
Alexia Dorffer, Samuel Wieczorek
}
