% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTest_plot.R
\name{hc_logFC_DensityPlot}
\alias{hc_logFC_DensityPlot}
\title{Density plots of logFC values}
\usage{
hc_logFC_DensityPlot(df_logFC, th_logFC = 0, pal = NULL)
}
\arguments{
\item{df_logFC}{A dataframe that contains the logFC values}

\item{th_logFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}

\item{pal}{xxx}
}
\value{
A highcharts density plot
}
\description{
This function show the density plots of Fold Change (the same as calculated
by limma) for a list of the comparisons of conditions in a differential
analysis.
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept
# Simulate missing value imputation
assay(obj[[1]])[which(is.na(assay(obj[[1]])))] <- 0
assay(obj[[2]])[which(is.na(assay(obj[[2]])))] <- 0

qData <- as.matrix(assay(obj[[2]]))
sTab <- SummarizedExperiment::colData(obj)
limma <- limmaCompleteTest(qData, sTab)
pal <- ExtendPalette(2, "Dark2")
hc_logFC_DensityPlot(limma$logFC, th_logFC = 1, pal = pal)

}
\author{
Samuel Wieczorek
}
