% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapForMissingValues}
\alias{heatmapForMissingValues}
\title{Display a heatmap for data with missing values}
\usage{
heatmapForMissingValues(
  x,
  col = NULL,
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{x}{A numeric matrix}

\item{col}{See \code{graphics::image()}}

\item{srtCol}{See \code{graphics::text()}}

\item{labCol}{See \code{graphics::text()}}

\item{labRow}{See \code{graphics::axis()}}

\item{key}{See \code{graphics::par()}}

\item{key.title}{See \code{graphics::title()}}

\item{main}{See \code{graphics::title()}}

\item{ylab}{See \code{graphics::image()}}
}
\value{
A heatmap
}
\description{
Display a heatmap for data with missing values
}
