% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{histPValue_HC}
\alias{histPValue_HC}
\title{Plots a histogram of p-values}
\usage{
histPValue_HC(pval_ll, bins = 80, pi0 = 1)
}
\arguments{
\item{pval_ll}{A vector of the p-values.}

\item{bins}{A integer indicating the number of cells for the histogram}

\item{pi0}{A \code{float} between 0 and 1 corresponding to the proportion of true null hypotheses.}
}
\value{
A histogram of the p-values with pi0
}
\description{
Plots a histogram of p-values
}
\examples{
data(subR25prot)
obj <- subR25prot
# Simulate imputation
obj <- NAIsZero(obj, 1)
obj <- NAIsZero(obj, 2)
allComp <- limmaCompleteTest(SummarizedExperiment::assay(obj[[length(obj)]]), design.qf(obj), comp.type="OnevsOne")
histPValue_HC(allComp$P_Value[1])

}
\author{
Samuel Wieczorek
}
