% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2_qfeatures.R
\name{createQFeatures}
\alias{createQFeatures}
\alias{createSampleinfo}
\title{Creates an object of class \code{QFeatures} from text file.}
\usage{
createQFeatures(
  data = NULL,
  file = "myDataset",
  sample,
  indQData,
  keyId = "AutoID",
  indexForMetacell = NULL,
  logData = FALSE,
  force.na = TRUE,
  typeDataset,
  parentProtId = NULL,
  analysis = "foo",
  description = NULL,
  processes = NULL,
  typePipeline = NULL,
  name.pipeline = NULL,
  software = NULL,
  name = "original"
)

createSampleinfo(samplename, conds, order = TRUE)
}
\arguments{
\item{data}{The name of a tab-separated file that contains the data.}

\item{file}{A \code{character()}.}

\item{sample}{A dataframe describing the samples (in lines).}

\item{indQData}{A vector of string where each element is the name
of a column in designTable that have to be integrated in
the \code{rowData()} table of the \code{QFeatures} object.}

\item{keyId}{A \code{character(1)} or \code{numeric(1)} which is the indice of the
column containing the ID of entities (peptides or proteins)}

\item{indexForMetacell}{They must be in the same order as the samples in
the experimental design}

\item{logData}{A \code{boolean} that indicates whether the data should be logged or not.}

\item{force.na}{A \code{boolean} that indicates if the '0' and 'NaN' values of
quantitative values  must be replaced by 'NA' (Default is FALSE)}

\item{typeDataset}{A string that indicates whether the dataset is about}

\item{parentProtId}{A \code{character(1)} For peptide entities, a string which
is the name of a column in rowData. It contains the id of parent proteins
and is used to generate adjacency matrix and process to aggregation.}

\item{analysis}{A \code{character()} which is the name of the MS study.}

\item{description}{A text which describes the study.}

\item{processes}{A vector of A \code{character()} which contains the name of
processes which has already been run on the data. Default is 'original'.}

\item{typePipeline}{A \code{character(1)} The type of pipeline used with this
dataset. The list of predefined pipelines in DaparToolshed can be obtained
with the function \code{pipelines()}. Default value is NULL}

\item{name.pipeline}{A string}

\item{software}{A \code{character()}}

\item{name}{A \code{character()} which is the name of the assay in the
QFeatures object. Default is 'original'}

\item{samplename}{A vector of string containing the names of the samples.}

\item{conds}{A vector of string containing the names of the conditions
associated to each sample. Must be the same length as samplename.}

\item{order}{Whether or not to order samples according to conditions.}
}
\value{
An instance of class \code{QFeatures}.

A \code{data.frame}.
}
\description{
Creates an object of class \code{QFeatures} from a
single tabulated-like file for quantitative and meta-data and a dataframe
for the samples description.

Creates a data.frame for sample information.
}
\examples{
NULL

NULL

}
\author{
Samuel Wieczorek, Manon Gaudin

Manon Gaudin
}
