% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infos_dataset.R
\name{infos_dataset}
\alias{infos_dataset}
\alias{infos_dataset_ui}
\alias{infos_dataset_server}
\title{infos_dataset_ui and infos_dataset_server}
\usage{
infos_dataset_ui(id)

infos_dataset_server(
  id,
  obj = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{shiny id}

\item{obj}{An instance of the class \code{QFeatures}.}
}
\value{
A shiny app
}
\description{
A shiny Module.
}
\examples{

library(QFeatures)
library(DaparViz)
library(shinyjs)
library(shiny)

data(ft_na, package='DaparViz')

ui <- fluidPage(infos_dataset_ui("mod_info"))

server <- function(input, output, session) {
  infos_dataset_server("mod_info", reactive({ft_na}))
}

shinyApp(ui, server)

}
\keyword{internal}
