% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.topn}
\alias{inner.aggregate.topn}
\title{xxxx}
\usage{
inner.aggregate.topn(pepData, X, method = "Mean", n = 10)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}

\item{method}{xxxxx}

\item{n}{xxxxx}
}
\value{
xxxxx
}
\description{
xxxx
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj[[length(obj)]])
inner.aggregate.topn(assay(obj[[length(obj)]]), X, n=3)
}

}
\author{
Samuel Wieczorek
}
