% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{inner.mean}
\alias{inner.mean}
\title{Mean-based aggregation}
\usage{
inner.mean(pepData, X)
}
\arguments{
\item{pepData}{A \code{matrix} containing the peptide intensities.}

\item{X}{A \code{matrix} acting as an adjacency matrix.}
}
\value{
A \code{matrix} containing the aggregated values.
}
\description{
Aggregation using mean method.
}
\examples{
\dontrun{
library(QFeatures)
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj)
i.mean <- inner.mean(assay(obj), X)
}

}
\author{
Samuel Wieczorek
}
