% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{inner.median}
\alias{inner.median}
\title{Median-based aggregation}
\usage{
inner.median(pepData, X)
}
\arguments{
\item{pepData}{A \code{matrix} containing the peptide intensities.}

\item{X}{A \code{matrix} acting as an adjacency matrix.}
}
\value{
A \code{matrix} containing the aggregated values.
}
\description{
Aggregation using median method.
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj)
i.mean <- inner.median(assay(obj), X)
}

}
\author{
Manon Gaudin
}
