% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{inner.robustsummary}
\alias{inner.robustsummary}
\title{robustSummary-based aggregation}
\usage{
inner.robustsummary(pepData, X)
}
\arguments{
\item{pepData}{A \code{matrix} containing the peptide intensities.
Note that the function assume that data is already log-transformed.}

\item{X}{A \code{matrix} acting as an adjacency matrix.}
}
\value{
A \code{matrix} containing the aggregated values.
}
\description{
Aggregation using robustSummary method.
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj)
i.mean <- inner.robustSummary(assay(obj), X)
}

}
\author{
Manon Gaudin
}
