% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.sum}
\alias{inner.sum}
\title{xxxx}
\usage{
inner.sum(pepData)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}
}
\value{
A matrix
}
\description{
xxxx
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(10)]
adjacencyMatrix(obj[[length(obj)]]) <- BuildAdjacencyMatrix(obj[[length(obj)]])
inner.sum(obj[[length(obj)]])
}
\author{
Samuel Wieczorek
}
