% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{is.differential}
\alias{is.differential}
\title{Identification of differentially abundant peptide/protein}
\usage{
is.differential(pvalue, logFC, thpvalue, thlogFC)
}
\arguments{
\item{pvalue}{A vector of p-values.}

\item{logFC}{A vector of logFC.}

\item{thpvalue}{A float indicating the p-value threshold.}

\item{thlogFC}{A float indicating the logFC threshold.}
}
\value{
A vector indicating which peptide/protein is differentially abundant (1) or not (0).
}
\description{
This function allows to identify differentially abundant peptide/protein
}
\examples{
data(subR25prot)
obj <- subR25prot
# Simulate imputation
obj <- NAIsZero(obj, 1)
obj <- NAIsZero(obj, 2)
allComp <- limmaCompleteTest(SummarizedExperiment::assay(obj[[length(obj)]]), design.qf(obj), comp.type="OnevsOne")
is.differential(allComp$P_Value[, 1], allComp$logFC[, 1], 0.05, 0.5)

}
\author{
Manon Gaudin
}
