% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.subset}
\alias{is.subset}
\title{Check is a given set is a subset of another one.}
\usage{
is.subset(set1, set2)
}
\arguments{
\item{set1}{A vector of \code{character()}}

\item{set2}{A vector of \code{character()}}
}
\value{
A \code{boolean}
}
\description{
Check is a given set is a subset of another one.
}
\examples{
is.subset('a', letters)
is.subset(c('a', 'c', 't'), letters)
is.subset(c('a', 3, 't'), letters)
is.subset(3, letters)

}
