% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isSubset}
\alias{isSubset}
\title{Check is a given set is a subset of another one.}
\usage{
isSubset(set1, set2)
}
\arguments{
\item{set1}{A vector of \code{character()}}

\item{set2}{A vector of \code{character()}}
}
\value{
A \code{boolean}
}
\description{
Check is a given set is a subset of another one.
}
\examples{
isSubset('a', letters)
isSubset(c('a', 'c', 't'), letters)
isSubset(c('a', 3, 't'), letters)
isSubset(3, letters)

}
