% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Keep_Items_from_Dataset}
\alias{Keep_Items_from_Dataset}
\title{Removes assay from a \code{QFeatures} object}
\usage{
Keep_Items_from_Dataset(dataset, range)
}
\arguments{
\item{dataset}{An instance of the \code{QFeatures} class}

\item{range}{A vector of integers}
}
\value{
The dataset minus some items
}
\description{
Removes one or more items from the dataset. This function is specific of the
type of dataset.
}
