% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepa.R
\name{LH0.lm}
\alias{LH0.lm}
\title{Compute the likelihoods of the null hypothesis for the "local" component-wise model}
\usage{
LH0.lm(X, y1, y2)
}
\arguments{
\item{X}{an n.pep*n.prot indicator matrix.}

\item{y1}{n.pep*n.samples matrice giving the observed counts for
each peptide in each sample from the condition 1}

\item{y2}{n.pep*n.samples matrice giving the observed counts for
each peptide in each sample from the condition 2}
}
\value{
A list of likelihoods with associated linear models
}
\description{
Compute the likelihoods of the null hypothesis for the "local" component-wise model
}
\examples{
NULL

}
\author{
Thomas Burger, Laurent Jacob
}
