% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepa.R
\name{LH0}
\alias{LH0}
\title{Compute the likelihood of the null hypothesis for the global model}
\usage{
LH0(X, y1, y2)
}
\arguments{
\item{X}{an n.pep*n.prot indicator matrix.}

\item{y1}{n.pep*n.samples matrice giving the observed counts for}

\item{y2}{n.pep*n.samples matrice giving the observed counts for}
}
\value{
A list of likelihoods
}
\description{
Compute the likelihood of the null hypothesis for the global model
}
\examples{
NULL
}
\author{
Thomas Burger, Laurent Jacob
}
