% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{makeDesign}
\alias{makeDesign}
\title{Builds the design matrix}
\usage{
makeDesign(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the \code{pData()} function
of package \code{MSnbase}.}
}
\value{
A design matrix
}
\description{
Builds the design matrix
}
\examples{
data(subR25pept)
makeDesign(SummarizedExperiment::colData(subR25pept))

}
\author{
Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek
}
