% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{makeDesign1}
\alias{makeDesign1}
\title{Builds the design matrix for designs of level 1}
\usage{
makeDesign1(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the \code{pData()} function
of package \code{MSnbase}.}
}
\value{
A design matrix
}
\description{
Builds the design matrix for designs of level 1
}
\examples{
data(subR25pept)
makeDesign1(SummarizedExperiment::colData(subR25pept))

}
\author{
Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek
}
