% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{makeDesign2}
\alias{makeDesign2}
\title{Builds the design matrix for designs of level 2}
\usage{
makeDesign2(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the \code{colData()} function
of package \code{MSnbase}.}
}
\value{
A design matrix
}
\description{
Builds the design matrix for designs of level 2
}
\examples{
data(subR25pept)
makeDesign2(SummarizedExperiment::colData(subR25pept))


}
\author{
Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek
}
