% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{makeDesign3}
\alias{makeDesign3}
\title{Builds the design matrix for designs of level 3}
\usage{
makeDesign3(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the \code{colData()} function
of package \code{MSnbase}.}
}
\value{
A design matrix
}
\description{
Builds the design matrix for designs of level 3
}
\examples{
data(subR25pept)
sTab <- cbind(SummarizedExperiment::colData(subR25pept), Tech.Rep = seq_len(6))
makeDesign3(sTab)


}
\author{
Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek
}
