% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{matchMetacell}
\alias{matchMetacell}
\title{Similar to the function \code{is.na()} but focused on the equality
with the paramter 'type'.}
\usage{
matchMetacell(metadata, pattern = NULL, level)
}
\arguments{
\item{metadata}{A data.frame}

\item{pattern}{The value to search in the dataframe}

\item{level}{A string designing the type of entity/pipeline.
Available values are: \code{peptide}, \code{protein}}
}
\value{
A boolean dataframe
}
\description{
Similar to the function \code{is.na()} but focused on the equality
with the paramter 'type'.
}
\examples{
data(subR25pept)
metadata <- qMetacell(subR25pept[[1]])
m <- matchMetacell(metadata, pattern = "Missing", level = "peptide")
m <- matchMetacell(metadata, pattern = 'Missing POV', level = "peptide")
m <- matchMetacell(metadata, pattern = c('Missing', 'Missing POV'), level = "peptide")
}
\author{
Samuel Wieczorek
}
