% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{metacell_agg}
\alias{metacell_agg}
\title{Metacell aggregation}
\usage{
metacell_agg(aggregatedSE, originalSE, adj_mat, conds, protname_order)
}
\arguments{
\item{aggregatedSE}{An instance of class \link{SummarizedExperiment} containing the aggregated data.}

\item{originalSE}{An instance of class \link{SummarizedExperiment} containing the non-aggregated data.}

\item{adj_mat}{An adjacency matrix.}

\item{conds}{xxx}

\item{protname_order}{xxx}
}
\value{
A \code{SummarizedExperiment} containing the aggregated data.
}
\description{
xxx
}
\author{
Samuel Wieczorek, Manon Gaudin
}
