% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Metacell_DIA_NN}
\alias{Metacell_DIA_NN}
\title{Sets the metacell dataframe for datasets which are from Dia-NN software}
\usage{
Metacell_DIA_NN(qdata, conds, df, level = NULL)
}
\arguments{
\item{qdata}{An object of class \code{MsnSet}}

\item{conds}{A 1-col dataframe with the condition associated to each sample.}

\item{df}{A dataframe with the same dimension as qdata containing the metacell.}

\item{level}{A string designing the type of entity/pipeline.
Available values are: \code{peptide}, \code{protein}}
}
\value{
NA
}
\description{
Actually, this function uses the generic function to generate metacell info
}
\examples{
data(subR25pept)
conds <- design.qf(subR25pept)$Condition
qdata <- SummarizedExperiment::assay(subR25pept[[2]])
df <- Metacell_DIA_NN(qdata, conds, df, level = "peptide")


}
\author{
Samuel Wieczorek
}
