% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_set_pval_threshold.R
\name{mod_set_pval_threshold}
\alias{mod_set_pval_threshold}
\alias{mod_set_pval_threshold_ui}
\alias{mod_set_pval_threshold_server}
\title{Module set pval threshold}
\usage{
mod_set_pval_threshold_ui(id)

mod_set_pval_threshold_server(
  id,
  pval_init = reactive({
     1
 }),
  fdr = reactive({
     NULL
 }),
  options = list(threshold = NULL)
)

mod_set_pval_threshold(xxx)
}
\arguments{
\item{id}{xxx}

\item{pval_init}{xxx}

\item{fdr}{xxx}

\item{options}{xxx}
}
\description{
A shiny Module.
}
