% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{mv_imputation_protein}
\alias{mv_imputation_protein}
\alias{findMECBlock}
\alias{reIntroduceMEC}
\alias{wrapper.impute.KNN}
\alias{wrapper.impute.fixedValue}
\alias{wrapper.impute.pa}
\alias{wrapper.impute.detQuant}
\alias{getQuantile4Imp}
\alias{wrapper.impute.slsa}
\title{Finds the LAPALA}
\usage{
findMECBlock(obj, grp)

reIntroduceMEC(obj, grp, MECIndex)

wrapper.impute.KNN(obj = NULL, grp, K)

wrapper.impute.fixedValue(obj, grp, fixVal = 0, na.type)

wrapper.impute.pa(obj = NULL, grp, q.min = 0.025)

wrapper.impute.detQuant(obj, qval = 0.025, factor = 1, na.type)

getQuantile4Imp(qdata, qval = 0.025, factor = 1)

wrapper.impute.slsa(obj = NULL, design = NULL)
}
\arguments{
\item{obj}{An object of class \code{QFeatures}.}

\item{grp}{A \code{vector} of conditions in the dataset}

\item{MECIndex}{A data.frame that contains index of MEC (see findMECBlock)}

\item{K}{the number of neighbors.}

\item{fixVal}{A float.}

\item{na.type}{A string which indicates the type of missing values to impute.
Available values are: \code{NA} (for both POV and MEC), \code{POV}, \code{MEC}.}

\item{q.min}{Same as the function \code{impute.pa()} in the package \code{imp4p}}

\item{qval}{An expression set containing quantitative values of various
replicates}

\item{factor}{A scaling factor to multiply the imputation value with}

\item{qdata}{A \code{data.frame()} for the quantitative data}

\item{design}{A \code{data.frame()} for the design of the dataset}
}
\value{
A data.frame containing the indexes of LAPALA

A list of two vectors, respectively containing the imputation values
and the rescaled imputation values
}
\description{
Methods available are:
\itemize{
\item wrapper.impute.detQuant():
This method is a wrapper of the function \code{impute.detQuant()} for objects
of class \code{MSnSet}
\item wrapper.impute.KNN(): Can impute only POV missing values. This method is
a wrapper for objects of class \code{QFeatures} and imputes missing values with
a fixed value. This function imputes the missing values condition by
condition.
\item wrapper.impute.slsa():
Imputation of peptides having no values in a biological condition. This
method is a wrapper to the function \code{impute.slsa()} of the package
\code{imp4p} adapted to an object of class \code{MSnSet}.
\item wrapper.impute.fixedValue():
This method is a wrapper to objects of class \code{MSnSet} and imputes
missing values with a fixed value.
\item wrapper.impute.pa():
Imputation of peptides having no values in a biological condition.
This method is a wrapper to the function \code{impute.pa} of the package
\code{imp4p} adapted to an object of class \code{MSnSet}.
}
}
\section{Utilities functions:}{
\itemize{
\item findMECBlock()
\item reIntroduceMEC()
\item getQuantile4Imp(): Quantile imputation value definition. This method
returns the q-th quantile of each column of an expression set, up to a s
calling factor
}
}

\examples{
data(subR25prot)
obj <- subR25prot[[2]]
grp <- design.qf(subR25prot)$Condition
lapala <- findMECBlock(obj, grp)
na.type = c("Missing POV", "Missing MEC")
obj.imp.pov <- wrapper.impute.detQuant(obj, na.type = na.type)
obj.imp.pov <- reIntroduceMEC(obj, grp, lapala)

obj.imp.pov <- wrapper.impute.KNN(obj, grp, 3)

obj.imp.pov <- wrapper.impute.fixedValue(obj, grp, 0.001, na.type = "Missing POV")
obj.imp.mec <- wrapper.impute.fixedValue(obj, grp, 0.001, na.type = "Missing MEC")
obj.imp.na <- wrapper.impute.fixedValue(
obj, grp, 0.001, 
na.type = c("Missing MEC", "Missing POV"))

obj.imp.pov <- wrapper.impute.pa(obj, grp)

qdata <- SummarizedExperiment::assay(obj)
quant <- getQuantile4Imp(qdata)



}
\author{
Samuel Wieczorek
}
