% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_QFeatures_Addons.R
\name{NAIsZero}
\alias{NAIsZero}
\title{Set NA values to 0}
\usage{
NAIsZero(obj, i)
}
\arguments{
\item{obj}{An instance of QFeatures class}

\item{i}{An integer which is the index of the assay in the QFeatures object}
}
\value{
An instance of QFeatures class
}
\description{
Set NA values to 0
}
