% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nEmptyLines}
\alias{nEmptyLines}
\title{Number of empty lines in the data}
\usage{
nEmptyLines(df)
}
\arguments{
\item{df}{A \code{data.frame}.}
}
\value{
A \code{integer(1)}
}
\description{
This function counts the number of empty lines (all elements
are equal to NA).
}
\details{
#' @title Version number of Prostar suite
#'
#' @description
#'
#' This function gives the version number of the packages of the
#' Prostar suite and which propose data processing. This information
#' can be useful if the user wants to publish its works or to
#' rerun a data processing pipeline tin a given set of conditions.
#' The packages which are concerned are \code{Prostar}, \code{DaparToolshed}
#' and \code{DaparToolshedData}
#'
#' @return A \code{list(3)}
#'
#' @rdname ProstarVersions
#'
#' @author Samuel Wieczorek
#'
#' @examples
#' ProstarVersions()
#'
#' @export
#'
ProstarVersions <- function() {
v.Prostar <- v.DaparToolshed  <- NA

\if{html}{\out{<div class="sourceCode">}}\preformatted{# tryCatch(
#     \{
#         find.package("Prostar.2.0")
#         v.Prostar <- Biobase::package.version("Prostar.2.0")
#     \},
#     error = function(e) v.Prostar <- NA
# )

tryCatch(
    \{
        find.package("DaparToolshed")
      v.DaparToolshed <- Biobase::package.version("DaparToolshed")
    \},
    error = function(e) v.DaparToolshed <- NA
)



list(
    Prostar = v.Prostar,
    DaparToolshed = v.DaparToolshed
)
}\if{html}{\out{</div>}}

}
}
\examples{
library(QFeatures)
data(subR25prot)
nEmptyLines(SummarizedExperiment::assay(subR25prot, 1))

}
\author{
Samuel Wieczorek
}
