% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_normalize.R
\name{normalizeFunction}
\alias{normalizeFunction}
\title{Normalisation for QFeatures}
\usage{
normalizeFunction(
  obj,
  method,
  conditions = NULL,
  type = "overall",
  subset.norm = NULL,
  quantile = 0.15,
  scaling = FALSE,
  span = 0.7
)
}
\arguments{
\item{obj}{An object of class \code{QFeatures} or \code{SummarizedExperiment}.
If data is of class \code{QFeatures}, the last assay will be normalized.}

\item{method}{Define the normalization method used : \verb{"GlobalQuantileAlignment"``,  }"QuantileCentering"\verb{, }"MeanCentering"\verb{, }"SumByColumns"\verb{, }"LOESS"\code{or}"vsn"`.}

\item{conditions}{A vector of conditions in the dataset.
If not provided, the vector \code{"Condition"} from the column metadata will be used.}

\item{type}{"overall" (shift all the sample distributions at once) or
"within conditions" (shift the sample distributions within each condition at a time).}

\item{subset.norm}{A vector of index indicating rows to be used for normalization}

\item{quantile}{A float that corresponds to the quantile used to align the data.}

\item{scaling}{A boolean that indicates if the variance of the data have to
be forced to unit (variance reduction) or not.}

\item{span}{xxx}
}
\value{
\code{QFeatures} including a new assay with normalized data or
\code{SummarizedExperiment} with normalized data.
}
\description{
This method is a wrapper that provides several methods to normalize quantitative
data from objects of class \code{QFeatures} or \code{SummarizedExperiment}.

They are organized in six main families : GlobalQuantileAlignement,
sumByColumns, QuantileCentering, MeanCentering, LOESS, vsn
For the first family, there is no type.
For the five other families, two type categories are available :
"Overall" which means that the value for each protein
(ie line in the expression data tab) is computed over all the samples ;
"within conditions" which means that the value for each protein
(ie line in the \code{SummarizedExperiment::assay()} data tab) is computed condition
by condition.
The available methods are described in \code{\link[=normalizeMethods]{normalizeMethods()}}.
}
\examples{
data(subR25pept)
normalized <- normalizeFunction(subR25pept, )

}
\author{
Manon Gaudin
}
