% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeMethods.dapar}
\alias{normalizeMethods.dapar}
\alias{normalizeMethods}
\title{List normalization methods with tracking option}
\usage{
normalizeMethods(target = "all")
}
\arguments{
\item{withTracking}{xxx}
}
\value{
xxx
}
\description{
List normalization methods with tracking option
}
\examples{
normalizeMethods()

}
