% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output2Excel.R
\name{output_2_Excel}
\alias{output_2_Excel}
\alias{readExcel}
\alias{listSheets}
\alias{write_Assay_To_Excel}
\alias{WriteHistory}
\alias{Write_SamplesData_to_Excel}
\alias{Write_RowData}
\alias{write.excel}
\title{This function exports a data.frame to a Excel file.}
\usage{
readExcel(file, sheet = NULL)

listSheets(file)

write_Assay_To_Excel(wb, obj, i, n)

WriteHistory(wb, obj, n)

Write_SamplesData_to_Excel(wb, obj, n)

Write_RowData(wb, obj, i, n)

write.excel(obj, filename)
}
\arguments{
\item{file}{The name of the Excel file.}

\item{sheet}{The name of the sheet}

\item{wb}{xxxx}

\item{obj}{An instance of the class \code{QFeatures}}

\item{i}{An integer which is the index of the assay in the QFeatures object}

\item{n}{xxx}

\item{filename}{A character string for the name of the Excel file.}
}
\value{
A Excel file (.xlsx)
}
\description{
This function exports a \code{MSnSet} data object to a Excel file.
Each of the three data.frames in the \code{MSnSet} object (ie experimental
data, phenoData and metaData are respectively integrated into separate sheets
in the Excel file).

The colored cells in the experimental data correspond to the original
missing values which have been imputed.
}
\examples{
\donttest{
library(QFeatures)
data(subR25prot)
df <- SummarizedExperiment::assay(subR25prot[[1]])
tags <- qMetacell(subR25prot[[1]])
colors <- list(
    "Missing POV" = "lightblue",
    "Missing MEC" = "orange",
    "Quant. by recovery" = "lightgrey",
    "Quant. by direct id" = "white",
    "Combined tags" = "red"
)
file <- tempfile('toto.xlsx')
write.excel(subR25prot, filename = file)
unlink(file)

data(subR25pept)
file <- tempfile('foo.xlsx')
write.excel(subR25pept, file)
unlink(file)
}


}
\author{
Samuel Wieczorek
}
