% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Parent}
\alias{Parent}
\title{Parent name of a node}
\usage{
Parent(level, node = NULL)
}
\arguments{
\item{level}{A string designing the type of entity/pipeline.
Available values are: \code{peptide}, \code{protein}}

\item{node}{A \code{character()}

#' @examples
Parent('protein', 'Missing')
Parent('protein', 'Missing POV')
Parent('protein', c('Missing POV', 'Missing MEC'))
Parent('protein', c('Missing', 'Missing POV', 'Missing MEC'))}
}
\value{
NA
}
\description{
Parent name of a node
}
