% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{Pipelines}
\alias{Pipelines}
\title{List of available pipelines in the package}
\usage{
Pipelines()

Pipelines()
}
\value{
NA

NA
}
\description{
Get the list of pipelines available in the package

Get the list of pipelines available in the package
}
