% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{plotJitter}
\alias{plotJitter}
\title{Jitter plot of CC}
\usage{
plotJitter(list.of.cc = NULL)
}
\arguments{
\item{list.of.cc}{List of cc such as returned by the function get.pep.prot.cc}
}
\value{
A plot
}
\description{
Jitter plot of CC
}
\examples{
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
ll <- get.pep.prot.cc(X)
plotJitter(ll)

}
\author{
Thomas Burger
}
