% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{pushpvalue}
\alias{pushpvalue}
\title{Push p-values based on metacell tags}
\usage{
pushpvalue(
  obj,
  pvalue,
  scope = "WholeMatrix",
  pattern = "Imputed MEC",
  percent = TRUE,
  threshold = 1,
  conditions = NULL,
  operator = ">=",
  level = NULL,
  value = 1.00000000001
)
}
\arguments{
\item{obj}{An object of class \code{QFeatures} or \code{SummarizedExperiment}.
If data is of class \code{QFeatures}, the last assay will be used.}

\item{pvalue}{A vector of p-values.}

\item{scope}{A string for scope to use.
Available values are "WholeLine", "WholeMatrix", "AllCond" and "AtLeastOneCond".}

\item{pattern}{A vector of tag to use.}

\item{percent}{A boolean to indicate whether the threshold represent an
absolute value (percent = FALSE) or a percentage (percent = TRUE).}

\item{threshold}{A value that corresponds to the threshold value.
Either an integer if percent = FALSE, or a float between 0 and 1 of percent = TRUE.}

\item{conditions}{A vector of conditions in the dataset.
If not provided, the vector \code{"Condition"} from the column metadata will be used.}

\item{operator}{A string for operator to use.
Available operators are "<=", "<", ">=", ">", "==" and "!=".}

\item{level}{A string for dataset type. Either "peptide" or "protein"
If not provided, the string obtained from \code{typeDataset(obj)} will be used.}

\item{value}{A float, value to assign to the pushed p-value.
By default, the value is set slightly above 1 to be able to differentiate the pushed value.}
}
\value{
A vector with pushed p-values.
}
\description{
This function allows to push p-values to 1 based on metacell tags.
}
\examples{
data(subR25prot)
obj <- subR25prot
# Simulate imputation
obj <- NAIsZero(obj, 1)
obj <- NAIsZero(obj, 2)
allComp <- limmaCompleteTest(SummarizedExperiment::assay(obj[[length(obj)]]), design.qf(obj), comp.type="OnevsOne")
pushpvalue(obj, allComp$P_Value[, 1], scope = "WholeMatrix", pattern = c("Missing MEC", "Missing POV"), percent = TRUE, threshold = 0.5, operator = ">=",)

}
\author{
Manon Gaudin
}
