% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed-filterFeaturesOneSE.R
\docType{class}
\name{QFeatures-filtering-oneSE}
\alias{QFeatures-filtering-oneSE}
\alias{filterFeaturesOneSE}
\alias{filterFeaturesOneSE,}
\alias{DaparToolsehd,}
\alias{FunctionFilter,}
\alias{VariableFilter}
\alias{FunctionFilter-class}
\alias{FunctionFilter}
\alias{filterFeaturesOneSE,QFeatures-method}
\title{Filter features of one SE based on their rowData}
\usage{
FunctionFilter(name, ...)

filterFeaturesOneSE(object, ...)

\S4method{filterFeaturesOneSE}{QFeatures}(object, i, name = "newAssay", filters)
}
\arguments{
\item{name}{A \code{character(1)} naming the new assay. Default is
\code{newAssay}. Note that the function will fail if there's
already an assay with \code{name}.}

\item{...}{Additional arguments}

\item{object}{An instance of class \code{QFeatures} or \code{SummarizedExperiment}.}

\item{i}{The index or name of the assay which features will be
filtered the create the new assay.}

\item{filters}{A \code{list()} containing instances of class \code{AnnotationFilter}
or \code{FunctionFilter}}
}
\value{
A filtered \code{QFeature} object
}
\description{
The \code{filterFeaturesOneSE} methods enables users to filter features
based on a variable in their \code{rowData}. It is directly inspired of the
function \code{filterFeature} of the package \code{QFeatures}.
The first difference is that the filter only applies to one
\code{SummarizedExperiment} contained in the object rather than applying on
all the SE.
This method generates a new \code{SummarizedExperiment} object which is added
to the \code{QFeatures} object. If the SE on which the filter applies is the
last one of the object, then a new SE is created If it is not the last one, the
new SE is added and all the further SE are deleted. The features matching
the. The filters can be provided as instances of class \code{AnnotationFilter}
(see the package \code{QFeatures}) or of class \code{FunctionFilter} (see below).
}
\section{Function filters}{


The function filters are filters as defined in the
\code{DaparToolshed} package. Each filter is defined by a name (which is the
name of a function) and a list which contains the parameters passed to the
function. Those filters can be created with the \code{FunctionFilter}
constructor.

Those functions are divided into two main categories:
\itemize{
\item the one that filter on one rowData feature,
\item the one based on a two-dimensional information such as the adjacency
matrix
}

for the first category, all filters of class \link[AnnotationFilter:AnnotationFilter]{AnnotationFilter} can be
used as they are used in \code{QFeatures}

For the second category, the package \code{DaparToolshed} provides filter
functions based either on the adjacency matrix:
\itemize{
\item \code{\link[=topnPeptides]{topnPeptides()}}
\item \code{\link[=sharedPeptides]{sharedPeptides()}}
\item \code{\link[=specPeptides]{specPeptides()}}
}

Or based on the quantitative metadata (identification):
\itemize{
\item \code{\link[=qMetacellWholeMatrix]{qMetacellWholeMatrix()}}
\item \code{\link[=qMetacellWholeLine]{qMetacellWholeLine()}}
\item \code{\link[=qMetacellOnConditions]{qMetacellOnConditions()}}
}
}

\examples{
data("subR25prot")
data("subR25pept")
## ----------------------------------------
## Creating function filters
## ----------------------------------------

#FunctionFilter('FUN',
#               param1 = 'value_of_param1',
#               param2 = 'value_of_param2')

FunctionFilter('qMetacellWholeLine',
               cmd = 'delete',
               pattern = 'Missing MEC')

## ----------------------------------------------------------------
## Filter the last assay to keep only specific peptides. This filter
## only applies on peptide dataset.
## ----------------------------------------------------------------

spec.filter <- FunctionFilter('specPeptides', list())
## using a user-defined character filter
filterFeaturesOneSE(subR25pept, filters = list(FunctionFilter('specPeptides', list())))


## ----------------------------------------------------------------
## Filter the last assay to keep only specific peptides and topn 
## peptides. The two filters are run sequentially.
## ----------------------------------------------------------------

lst.filters <- list(FunctionFilter('specPeptides', list()))
lst.filters <- append(lst.filters,
FunctionFilter('topnPeptides',
fun = 'rowSums',
top = 2))
filterFeaturesOneSE(subR25pept, filters = lst.filters)

## ----------------------------------------------------------------
## Filter the last assay to delete peptides where, in at least one 
## condition, there is less than 80\% of samples marked as 'imputed POV'
## ----------------------------------------------------------------

filter <- FunctionFilter('qMetacellOnConditions',
cmd = 'delete',
mode = 'AtLeastOneCond',
pattern = 'Missing POV',
conds = SummarizedExperiment::colData(subR25prot)$Condition,
percent = TRUE,
th = 0.8,
operator = '<')

 filterFeaturesOneSE(subR25prot, filters = list(filter))


}
\author{
Samuel Wieczorek
}
