% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFeatures_utils.R
\name{QFeatures-utils}
\alias{QFeatures-utils}
\alias{last_assay}
\alias{n_assays_in_qf}
\alias{QFeaturesFromSE}
\title{Utility funcitons to dela with QFeatures objects.}
\usage{
last_assay(object)

n_assays_in_qf(object)

QFeaturesFromSE(
  obj.se,
  colData = data.frame(),
  metadata.qf = data.frame(),
  name = "myname"
)
}
\arguments{
\item{object}{An instance of the class \code{QFeatures}}

\item{obj.se}{An instance of the class \code{QFeatures}}

\item{colData}{A data.frame() which contains data for the future object,}

\item{metadata.qf}{A data.frame() which contains the metadata for the future object,}

\item{name}{The name of the assay in the QFeatures object}
}
\value{
NA

An instance of QFeatures class
}
\description{
Utility funcitons to dela with QFeatures objects.
}
\examples{

NULL

# example code

}
