% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_query_metacell.R
\name{query-metacell}
\alias{query-metacell}
\alias{mod_query_metacell_ui}
\alias{mod_query_metacell_server}
\alias{mod_query_metacell}
\title{Query cell metadata module}
\usage{
mod_query_metacell_ui(id)

mod_query_metacell_server(
  id,
  obj = reactive({
     NULL
 }),
  filters = reactive({
     NULL
 }),
  val_vs_percent = reactive({
     NULL
 }),
  operator = reactive({
     NULL
 }),
  reset = reactive({
     NULL
 }),
  op_names = reactive({
     NULL
 })
)

mod_query_metacell(obj)
}
\arguments{
\item{id}{A character(1)}

\item{obj}{An instance of the class \code{SummarizedExperiment}}

\item{filters}{Available values are}

\item{val_vs_percent}{Available values are}

\item{operator}{Available values are}

\item{reset}{xxx}

\item{op_names}{Available values are}
}
\value{
NA
}
\description{
xxxx
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(Exp1_R25_prot, package = 'DaparToolshedData')
shiny::runApp(mod_query_metacell(Exp1_R25_prot[[1]]))

\dontshow{\}) # examplesIf}
}
