% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{RunAggregation}
\alias{RunAggregation}
\title{Aggregation of peptide-level assay QFeatures}
\usage{
RunAggregation(
  qf,
  includeSharedPeptides = "Yes_As_Specific",
  operator = "Mean",
  considerPeptides = "allPeptides",
  adjMatrix = "adjacencyMatrix",
  ponderation = "Global",
  n = NULL,
  aggregated_col = NULL,
  max_iter = 500
)
}
\arguments{
\item{qf}{An instance of class \link[QFeatures:QFeatures]{QFeatures}. The last assay contained in \code{qf} will be aggregated. Intensities are assumed to already be log-transformed.}

\item{includeSharedPeptides}{How shared peptides are handled. Either \code{Yes_As_Specific} (default), \code{Yes_Iterative_Redistribution}, \code{Yes_Simple_Redistribution} or \code{No}.
See below for details.}

\item{operator}{A function used for quantitative feature aggregation.
Available functions are \code{Sum}, \code{Mean}, \code{Median}, \code{medianPolish} or \code{robustSummary}.
See below for details.}

\item{considerPeptides}{A \code{character(1)} defining what peptide to consider. Available values are \code{allPeptides} (default) and \code{topN}.}

\item{adjMatrix}{A \code{character(1)} naming a rowdata variable from the last assay of \code{qf} containing an adjacency matrix.}

\item{ponderation}{A \code{character(1)} defining what to consider to create the coefficient for redistribution of shared peptides.
Available values are \code{Global} (default), \code{Condition} or \code{Sample}.}

\item{n}{If \code{topN}, specify the number of peptides to use for each protein.}

\item{aggregated_col}{A \code{character()} of column names from rowdata to be aggregated.}

\item{max_iter}{A \code{numeric(1)} setting the maximum number of iteration.}
}
\value{
A QFeatures with an aggregated assay added.
}
\description{
This function aggregate both quantitative and rowdata from the last assay contained in a \code{QFeatures}.
Note that the function assumes that the intensities in the QFeatures are already log-transformed.
}
\details{
Aggregation of quantitative data is performed using aggregateFeatures, or
innerAggregateIter if \code{Yes_Iterative_Redistribution} or \code{Yes_Simple_Redistribution}
is selected.

The handling of shared peptide is as follow :
\itemize{
\item \code{Yes_As_Specific} : Shared peptides are used multiple times.
Each peptide is duplicated as many times as the number of proteins in
which they are present, and thus are considered as if they are specific
to each protein.
\item \code{Yes_Simple_Redistribution} : Intensity of shared peptides are
redistributed proportionally to each protein. See \code{innerAggregateIter}
for more information.
\item \code{Yes_Iterative_Redistribution} : Intensity of shared peptides are
redistributed proportionally to each protein. See \code{innerAggregateIter}
for more information.
\item \code{No} : No shared peptides are used. If a peptide contained only shared
peptides, its intensity is set as 0 for every sample.
}

Available functions are :
\itemize{
\item \code{Sum} : base::colSums()] or base::rowSums() if \code{Yes_Iterative_Redistribution} or \code{Yes_Simple_Redistribution}.
\item \code{Mean} : base::colMeans()] or base::rowMeans() if \code{Yes_Iterative_Redistribution} or \code{Yes_Simple_Redistribution}.
\item \code{Median} : matrixStats::mcolMedians()] or matrixStats::rowMedians() if \code{Yes_Iterative_Redistribution} or \code{Yes_Simple_Redistribution}.
\item \code{medianPolish} : MsCoreUtils::medianPolish().
\item \code{robustSummary} : MsCoreUtils::robustSummary().
}
}
\examples{
data(subR25pept)

# Remove empty lines
filter_emptyline <- FunctionFilter("qMetacellWholeLine", 
cmd = 'delete', pattern = 'Missing MEC')
subR25pept <- filterFeaturesOneSE(object = subR25pept, 
i = length(subR25pept), name = "Filtered",
              filters = list(filter_emptyline))
# Remove proteins with no peptide associated in adjacency matrix
indx <- which(Matrix::colSums(
SummarizedExperiment::rowData(
subR25pept[[length(subR25pept)]])$adjacencyMatrix) != 0)
SummarizedExperiment::rowData(
subR25pept[[length(subR25pept)]])$adjacencyMatrix <- 
SummarizedExperiment::rowData(
subR25pept[[length(subR25pept)]])$adjacencyMatrix[, indx]
  
obj.agg <- RunAggregation(subR25pept, "Yes_As_Specific", "Sum", "allPeptides",
aggregated_col = c("Sequence", "Mass"))
obj.agg <- RunAggregation(subR25pept, "Yes_As_Specific", "Mean", "allPeptides",
aggregated_col = c("Sequence", "Mass"))
obj.agg <- RunAggregation(subR25pept, "Yes_As_Specific", "Sum", "topN", n = 4,
aggregated_col = c("Sequence", "Mass"))
obj.agg <- RunAggregation(subR25pept, "Yes_As_Specific", "Mean", "topN", n = 4,
aggregated_col = c("Sequence", "Mass"))

obj.agg <- RunAggregation(subR25pept, "No", "Sum", "allPeptides")
obj.agg <- RunAggregation(subR25pept, "No", "Sum", "topN", n = 4)

obj.agg <- RunAggregation(subR25pept, "Yes_Simple_Redistribution", "Sum", "allPeptides",
aggregated_col = c("Sequence", "Mass"))
obj.agg <- RunAggregation(subR25pept, "Yes_Iterative_Redistribution", "Sum", "topN", n = 4,
aggregated_col = c("Sequence", "Mass"))

}
\author{
Samuel Wieczorek, Manon Gaudin
}
