% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{select_topn}
\alias{select_topn}
\title{Selection of top n peptides}
\usage{
select_topn(pepData, X, n = 10, funpept = "Mean")
}
\arguments{
\item{pepData}{A \code{matrix} containing the peptide intensities.}

\item{X}{A \code{matrix} acting as an adjacency matrix.}

\item{n}{A \code{numeric(1)} specifying the number of peptides to use for each protein.}

\item{funpept}{A function used for determining a peptide's value.
Available functions are \code{Sum}, \code{Mean} or \code{Median}.}
}
\value{
An adjacency matrix with only the top n peptides selected.
}
\description{
xxx
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(50)]
X <- BuildAdjacencyMatrix(obj[[length(obj)]])
X.topn <- select_topn(assay(obj[[length(obj)]]), X, n = 3)
}
}
\author{
Manon Gaudin
}
