% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatrix.R
\name{splitAdjacencyMat}
\alias{splitAdjacencyMat}
\title{splits an adjacency matrix into specific and shared}
\usage{
splitAdjacencyMat(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of two adjacency matrices
}
\description{
Method to split an adjacency matrix into specific and shared
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
obj.pep <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj.pep[[length(obj.pep)]])
ll <- splitAdjacencyMat(X)

}
\author{
Samuel Wieczorek
}
