% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{test.design}
\alias{test.design}
\title{Check the validity of the design}
\usage{
test.design(tab)
}
\arguments{
\item{tab}{A data.frame which correspond to the design}
}
\value{
A list of two items
}
\description{
Check the validity of the design
}
\examples{
data(subR25pept)
test.design(SummarizedExperiment::colData(subR25pept)[, seq(3)])

}
\author{
Thomas Burger, Samuel Wieczorek
}
