% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{testDesign}
\alias{testDesign}
\title{Check the validity of the design}
\usage{
testDesign(tab)
}
\arguments{
\item{tab}{A data.frame which correspond to the design}
}
\value{
A list of two items
}
\description{
Check the validity of the design
}
\examples{
data(subR25pept)
testDesign(SummarizedExperiment::colData(subR25pept)[, -1])

}
\author{
Thomas Burger, Samuel Wieczorek
}
