% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{thresholdpval4fdr}
\alias{thresholdpval4fdr}
\title{Applies p-value adjustment using cp4p on p-values below a given threshold}
\usage{
thresholdpval4fdr(x, pval.T, M)
}
\arguments{
\item{x}{vector of p-values}

\item{pval.T}{p-value threshold}

\item{M}{a pi0.method from cp4p package}
}
\value{
adjusted p-values
}
\description{
Applies p-value adjustment using cp4p on p-values below a given threshold
}
\examples{
NULL


}
\author{
Thomas Burger
}
