% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\alias{mod_volcanoplot_ui}
\alias{mod_volcanoplot_server}
\title{popover_for_help_ui and popover_for_help_server}
\usage{
mod_volcanoplot_ui(id)

mod_volcanoplot_server(
  id,
  data = reactive({
     NULL
 }),
  comp = reactive({
     NULL
 }),
  tooltip = reactive({
     NULL
 }),
  reset = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 })
)

volcanoplot(xxx)
}
\arguments{
\item{id}{xxx}

\item{data}{xxx}

\item{comp}{xxx}

\item{tooltip}{xxx}

\item{reset}{xxx}

\item{is.enabled}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

shiny::runApp(volcanoplot(xxxx))
\dontshow{\}) # examplesIf}
}
