% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_PeptideLevel.R
\name{wrapper.impute.mle}
\alias{wrapper.impute.mle}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
wrapper.impute.mle(obj, grp)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment}.}

\item{grp}{A vector of conditions in the dataset.}
}
\value{
The \code{SummarizedExperiment::assay(obj)} matrix with imputed values instead
of missing values.
}
\description{
This method is a wrapper to the function \code{impute.mle()} of the
package \code{imp4p} adapted to an object of class \code{SummarizedExperiment}.
It does not impute MEC missing values.
}
\examples{
utils::data(subR25pept)
level <- 'peptide'
# Delete whole empty lines
metacell.mask <- DaparToolshed::match.metacell(
qMetacell(subR25pept[[1]]), 
c("Missing POV", "Missing MEC"), level)
indices <- GetIndices_WholeMatrix(metacell.mask, op = ">=", th = 1)
grp <- design.qf(subR25pept)$Condition
subR25pept <- wrapper.impute.mle(subR25pept[[2]], grp)

}
\author{
Samuel Wieczorek
}
