% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_PeptideLevel.R
\name{wrapper.pirat}
\alias{wrapper.pirat}
\title{Missing values imputation using Pirat}
\usage{
wrapper.pirat(data, adjmat, rnas_ab = NULL, adj_rna_pg = NULL, ...)
}
\arguments{
\item{data}{An object of class \code{QFeatures} or \code{SummarizedExperiment}.
If data is of class \code{QFeatures}, the last assay will be imputed.}

\item{adjmat}{Adjacency matrix corresponding to the \code{SummarizedExperiment} or the last assay of \code{QFeatures}.}

\item{rnas_ab}{Transcriptomic data with sample as row, used only if extension = 'T'.}

\item{adj_rna_pg}{Adjacency matrix of rna (rows) and peptides or precursors (columns), used only if extension = 'T'.}

\item{...}{Additional arguments to pass to \code{my_pipeline_llkimpute()}}
}
\value{
\code{QFeatures} including a new assay with imputed data or \code{SummarizedExperiment} with imputed data.
}
\description{
This method is a wrapper to the function \code{pipeline_llkimpute()} of the
package \code{Pirat} adapted to an object of class \code{QFeatures} of \code{SummarizedExperiment}.
}
\examples{
data(subR25pept)

# Delete whole empty lines
filter_emptyline <- FunctionFilter("qMetacellWholeLine", cmd = 'delete', pattern = 'Missing MEC')
subR25pept <- filterFeaturesOneSE(object = subR25pept, i = length(subR25pept), name = "Filtered", 
              filters = list(filter_emptyline))

subR25pept <- wrapper.pirat(data = subR25pept, i = length(subR25pept), extension = "base")

}
\author{
Manon Gaudin
}
