% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{wrapperClassic1wayAnova}
\alias{wrapperClassic1wayAnova}
\title{Wrapper for One-way Anova statistical test}
\usage{
wrapperClassic1wayAnova(obj, i, with_post_hoc = "No", post_hoc_test = "No")
}
\arguments{
\item{obj}{An object of class \code{QFeatures}.}

\item{i}{An integer which is the index of the assay in the QFeatures object}

\item{with_post_hoc}{a character string with 2 possible values: "Yes" and
"No" (default) saying if function must perform a Post-Hoc test or not.}

\item{post_hoc_test}{character string, possible values are "No" (for no
test; default value) or TukeyHSD" or "Dunnett". See details of
\code{postHocTest()} function to choose the appropriate one.}
}
\value{
A list of two dataframes. First one called "logFC" contains
all pairwise comparisons logFC values (one column for one comparison) for
each analysed feature (Except in the case without post-hoc testing, for
which NAs are returned.); The second one named "P_Value" contains the
corresponding p-values.
}
\description{
Wrapper for One-way Anova statistical test
}
\details{
This function allows to perform a 1-way Analysis of Variance. Also
computes the post-hoc tests if the \code{with_post_hoc} parameter is set to
yes. There are two possible post-hoc tests: the Tukey Honest Significant
Differences (specified as "TukeyHSD") and the Dunnett test
(specified as "Dunnett").
}
\examples{
\donttest{
library(SummarizedExperiment)
data(subR25prot)
obj <- subR25prot
filter <- FunctionFilter('qMetacellOnConditions',
  cmd = 'delete',
  mode = 'AtLeastOneCond',
  pattern = c("Missing POV", "Missing MEC"),
  conds = design_qf(obj)$Condition,
  percent = TRUE,
  th = 0.65,
  operator = '>')
obj <- filterFeaturesOneSE(obj, name = "Filtered", filters = list(filter))
anovatest <- wrapperClassic1wayAnova(obj, 2)
}

}
\seealso{
postHocTest()
}
\author{
Hélène Borges
}
