% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{Add_Aggregated_rowData}
\alias{Add_Aggregated_rowData}
\title{Add aggregated rowData}
\usage{
Add_Aggregated_rowData(obj, col, i.agg)
}
\arguments{
\item{obj}{An instance of class \link[QFeatures:QFeatures-class]{QFeatures::QFeatures}.}

\item{col}{A \code{character()} of column names from rowdata to be aggregated.}

\item{i.agg}{A \code{numeric(1)} indicating the index of the assay to which add the aggregated rowData, using the previous assay's rowData.}
}
\value{
An instance of \code{QFeatures} class with aggregated rowData in specified assay.
}
\description{
Aggregation of rowData of a \code{QFeatures} assay.
}
\author{
Samuel Wieczorek, Manon Gaudin
}
