% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{applyAnovasOnProteins}
\alias{applyAnovasOnProteins}
\title{iteratively applies OWAnova() on the features of an MSnSet object}
\usage{
applyAnovasOnProteins(obj, i)
}
\arguments{
\item{obj}{a QFreatures object}

\item{i}{xxx
'}
}
\value{
a list of linear models
}
\description{
iteratively applies OWAnova() on the features of an MSnSet object
}
\examples{
data(Exp1_R25_prot, package='DaparToolshedData')
exdata <- Exp1_R25_prot[1:5,]
applyAnovasOnProteins(exdata, 1)

}
\author{
Thomas Burger
}
