% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatrix.R
\name{BuildAdjacencyMatrix}
\alias{BuildAdjacencyMatrix}
\title{Function matrix of appartenance group}
\usage{
BuildAdjacencyMatrix(obj.pep)
}
\arguments{
\item{obj.pep}{An object (peptides) of class \code{MSnSet}.}
}
\value{
A binary matrix
}
\description{
Method to create a binary matrix with proteins in columns and peptides
in lines on a \code{MSnSet} object (peptides)
}
\examples{
data(subR25pept)
BuildAdjacencyMatrix(subR25pept[[1]])

}
\author{
Florence Combes, Samuel Wieczorek, Alexia Dorffer
}
