% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{BuildMetacell}
\alias{BuildMetacell}
\title{metacell function which xxx}
\usage{
BuildMetacell(from = NULL, level, qdata = NULL, conds = NULL, df = NULL)
}
\arguments{
\item{from}{xxx}

\item{level}{xxx}

\item{qdata}{A matrix of quantitative data}

\item{conds}{xxx}

\item{df}{A data.frame which contains the type of identification of the
entities. It must have the same dimensions as \code{qData}.}
}
\value{
xxxxx
}
\description{
xxx
}
\examples{
data(subR25pept)
conds <- design.qf(subR25pept)$Condition
qdata <- SummarizedExperiment::assay(subR25pept[[2]])
metacell1 <- BuildMetacell('maxquant', 'peptide', qdata, conds)
metacell2 <- BuildMetacell('proline', 'peptide', qdata, conds)
}
\author{
Samuel Wieczorek
}
