% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{check.conditions}
\alias{check.conditions}
\title{Check if the design is valid}
\usage{
check.conditions(conds)
}
\arguments{
\item{conds}{A \code{vector} containing the conditions.}
}
\value{
A \code{list} including :
"valid" : Wether the conditions are valid or not.
"warn" : A message describing the issue if the conditions ar not valid.
}
\description{
Check if the design is valid
}
\examples{
data(subR25pept)
check.conditions(design.qf(subR25pept)$Condition)

}
\author{
Samuel Wieczorek
}
