% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Children}
\alias{Children}
\title{Names of all chidren of a node}
\usage{
Children(level, parent = NULL)
}
\arguments{
\item{level}{A string designing the type of entity/pipeline.
Available values are: \code{peptide}, \code{protein}}

\item{parent}{xxx}
}
\value{
A vector
}
\description{
xxx
}
\examples{
Children('protein', 'Missing')
Children('protein', 'Missing POV')
Children('protein', c('Missing POV', 'Missing MEC'))
Children('protein', c('Missing', 'Missing POV', 'Missing MEC'))
}
