% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{classic1wayAnova}
\alias{classic1wayAnova}
\title{Function to perform a One-way Anova statistical test on a MsnBase
dataset}
\usage{
classic1wayAnova(current_line, conditions)
}
\arguments{
\item{current_line}{The line currently treated from the quantitative data to
perform the ANOVA}

\item{conditions}{The conditions represent the different classes of the
studied factor}
}
\value{
A named vector containing all the different values of the aov model
}
\description{
Function to perform a One-way Anova statistical test on a MsnBase
dataset
}
\examples{
\donttest{
library(SummarizedExperiment)
data(subR25prot)
obj <- subR25prot
filter <- FunctionFilter('qMetacellOnConditions',
  cmd = 'delete',
  mode = 'AtLeastOneCond',
  pattern = c("Missing POV", "Missing MEC"),
  conds = design.qf(obj)$Condition,
  percent = TRUE,
  th = 0.8,
  operator = '>')
obj <- filterFeaturesOneSE(obj, name = "Filtered", filters = list(filter))

qdata <- SummarizedExperiment::assay(obj[[3]])
conds <- design.qf(obj)$Condition
anova_tests <- apply(qdata, 1, classic1wayAnova, conditions = as.factor(conds))
anova_tests <- t(anova_tests)

}

}
\author{
Hélène Borges
}
