% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CleanRowData}
\alias{CleanRowData}
\title{Customised resetZoomButton of highcharts plots}
\usage{
CleanRowData(obj, i)
}
\arguments{
\item{obj}{An instance of the class \code{QFeatures}}

\item{i}{An integer which is the index of the assay in the QFeatures object}
}
\value{
An instance of the SummarizedExperiment structure
}
\description{
Customised resetZoomButton of highcharts plots
}
\examples{
library("highcharter")
hc <- highchart()
hc_chart(hc, type = "line")
hc_add_series(hc, data = c(29, 71, 40))
my_hc_ExportMenu(hc, filename = "foo")

}
\author{
Samuel Wieczorek
}
