% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CleanRowData}
\alias{CleanRowData}
\title{Customised resetZoomButton of highcharts plots}
\usage{
CleanRowData(obj)
}
\arguments{
\item{obj}{xxx}
}
\description{
Customised resetZoomButton of highcharts plots
}
\examples{
library("highcharter")
hc <- highchart()
hc_chart(hc, type = "line")
hc_add_series(hc, data = c(29, 71, 40))
my_hc_ExportMenu(hc, filename = "foo")

}
\author{
Samuel Wieczorek
}
