% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{CountPep}
\alias{CountPep}
\title{Compute the number of peptides used to aggregate proteins}
\usage{
CountPep(X)
}
\arguments{
\item{X}{A "valued" adjacency matrix in which lines and columns correspond
respectively to peptides and proteins.}
}
\value{
A vector of boolean which is the adjacency matrix
but with NA values if they exist in the intensity matrix.
}
\description{
This function computes the number of peptides used to aggregate proteins.
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj.pep <- Exp1_R25_pept[seq_len(10)]
last.obj <- obj.pep[[length(obj.pep)]]
X <- BuildAdjacencyMatrix(last.obj)
CountPep(X)
}

}
\author{
Alexia Dorffer
}
