% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{ExtractUniquePeptides}
\alias{ExtractUniquePeptides}
\title{Get unique peptide}
\usage{
ExtractUniquePeptides(X)
}
\arguments{
\item{X}{A \code{matrix} acting as an adjacency matrix.}
}
\description{
Get unique peptide
}
\examples{
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[1]])
ExtractUniquePeptides(X)

}
