% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{ExtractUniquePeptides}
\alias{ExtractUniquePeptides}
\title{Test}
\usage{
ExtractUniquePeptides(X)
}
\arguments{
\item{X}{xxx}
}
\description{
Test
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(10)]
last.obj <- Exp1_R25_pept[[length(Exp1_R25_pept)]]
X <- BuildAdjacencyMatrix(last.obj)
ExtractUniquePeptides(X)
}

}
