% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{ExtractUniquePeptides}
\alias{ExtractUniquePeptides}
\title{Test}
\usage{
ExtractUniquePeptides(X)
}
\arguments{
\item{X}{xxx}
}
\description{
Test
}
\examples{
#' data(Exp1_R25_pept, package="DaparToolshedData")
X <- BuildAdjacencyMatrix(Exp1_R25_pept[seq_len(10)])
ExtractUniquePeptides(X)

}
