% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{formatPHResults}
\alias{formatPHResults}
\title{Extract logFC and raw pvalues from multiple post-hoc models summaries}
\usage{
formatPHResults(post_hoc_models_summaries)
}
\arguments{
\item{post_hoc_models_summaries}{a list of summaries of post-hoc models.}
}
\value{
a list of 2 dataframes containing the logFC values and pvalues for
each comparison.
}
\description{
Extract logFC and raw pvalues from multiple post-hoc models summaries
}
\examples{
\donttest{

library(SummarizedExperiment)
data(subR25prot)
obj <- subR25prot

obj <- NAIsZero(obj, 1)
obj <- NAIsZero(obj, 2)
qdata <- SummarizedExperiment::assay(obj[[2]])
conds <- design.qf(obj)$Condition
anova_tests <- apply(qdata, 1, classic1wayAnova, conditions = as.factor(conds))
anova_tests <- t(anova_tests)

names(anova_tests) <- rownames(qdata)
tms <- lapply(
  anova_tests,
  function(x) {
    summary(multcomp::glht(x,
      linfct = multcomp::mcp(conditions = "Tukey")
    ),
      test = multcomp::adjusted("none")
    )
   }
)
res <- formatPHResults(tms)
}

}
\author{
Helene Borges
}
