% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{getDesignLevel}
\alias{getDesignLevel}
\title{Get design level}
\usage{
getDesignLevel(sTab)
}
\arguments{
\item{sTab}{A \code{data.frame()} which contains the sample data}
}
\value{
An integer
}
\description{
Get design level
}
\examples{
data(subR25pept)
sTab <- SummarizedExperiment::colData(subR25pept)
getDesignLevel(sTab)

}
