% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_FunctionFiltering.R
\name{GetIndices_FunFiltering}
\alias{GetIndices_FunFiltering}
\title{Delete the lines in the matrix of intensities and the metadata table
given their indices.}
\usage{
GetIndices_FunFiltering(
  obj,
  conds,
  level,
  pattern = NULL,
  type = NULL,
  percent,
  op,
  th
)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment} containing
quantitative data.}

\item{conds}{A vector containing the names of the conditions from the sample.}

\item{level}{A vector of integers which are the indices of lines to
delete.}

\item{pattern}{A string to be included in the \code{SummarizedExperiment}
object for log.}

\item{type}{Available values are:
\itemize{
\item 'AllCond' (the query is valid in all the conditions),
\item 'AtLeatOneCond' (the query is valid in at leat one condition.
}}

\item{percent}{A boolean to indicate whether the threshold represent an
absolute value (percent = FALSE) or
a percentage (percent=TRUE).}

\item{op}{String for operator to use. List of operators is available with
'SymFilteringOperators()'.}

\item{th}{A floating number which is in the interval \verb{[0, 1]}}
}
\value{
An instance of class \code{SummarizedExperiment} that have been filtered.
}
\description{
Delete the lines in the matrix of intensities and the metadata table
given their indices.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
NA
\dontshow{\}) # examplesIf}
}
\author{
Samuel Wieczorek
}
