% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DAPAR.R, R/utils.R
\name{getListNbValuesInLines}
\alias{getListNbValuesInLines}
\title{Returns the possible number of values in lines in the data}
\usage{
getListNbValuesInLines(object, conds, type = "WholeMatrix")

getListNbValuesInLines(object, conds, type = "WholeMatrix")
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{conds}{xxxx}

\item{type}{WholeMatrix, AllCond or AtLeastOneCond}

\item{obj}{An object of class \code{MSnSet}}
}
\value{
An integer

An integer
}
\description{
Returns the possible number of values in lines in the data

Returns the possible number of values in lines in the data
}
\examples{
data(Exp1_R25_pept, package="DAPARdata")
getListNbValuesInLines(Exp1_R25_pept, "WholeMatrix")

data(ft, package='DaparToolshed')
res <- getListNbValuesInLines(ft, 1)

}
\author{
Samuel Wieczorek

Samuel Wieczorek, Enora Fremy
}
