% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GetNbPeptidesUsed}
\alias{GetNbPeptidesUsed}
\title{Computes the number of peptides used for aggregating each protein}
\usage{
GetNbPeptidesUsed(pepData)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A data.frame
}
\description{
Method to compute the number of quantified peptides used for aggregating
each protein
}
\examples{
library(QFeatures)
data(Exp1_R25_pept, package="DaparToolshedData")
obj.pep <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj.pep[[length(obj.pep)]])
adjacencyMatrix(obj.pep[[length(obj.pep)]]) <- X
GetNbPeptidesUsed(obj.pep)

}
\author{
Samuel Wieczorek
}
