% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{GetNbTags}
\alias{GetNbTags}
\title{Number of each metacell tags}
\usage{
GetNbTags(obj)
}
\arguments{
\item{obj}{A instance of the class \code{SummarizedExperiment}}
}
\description{
Number of each metacell tags
}
\examples{
data(Exp1_R25_prot, package = 'DaparToolshedData')
GetNbTags(Exp1_R25_prot[[1]])

}
