% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{getProteinsStats}
\alias{getProteinsStats}
\title{Computes the number of proteins that are only defined by
specific peptides, shared peptides or a mixture of two.}
\usage{
getProteinsStats(X)
}
\arguments{
\item{X}{The adjacency matrix with both specific and
shared peptides.}
}
\value{
A list
}
\description{
This function computes the number of proteins that are only defined by
specific peptides, shared peptides or a mixture of two.
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(20)]
obj.last <- obj[[length(obj)]]
X <- BuildAdjacencyMatrix(obj.last)
getProteinsStats(X)
}

}
\author{
Samuel Wieczorek
}
