% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTest_plot.R
\name{hc_logFC_DensityPlot}
\alias{hc_logFC_DensityPlot}
\title{Density plots of logFC values}
\usage{
hc_logFC_DensityPlot(df_logFC, th_logFC = 0, pal = NULL)
}
\arguments{
\item{df_logFC}{A dataframe that contains the logFC values}

\item{th_logFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}

\item{pal}{A \code{vector} of HEX codes for colors.}
}
\value{
A highcharts density plot
}
\description{
This function show the density plots of Fold Change (the same as calculated
by limma) for a list of the comparisons of conditions in a differential
analysis.
}
\examples{
library(SummarizedExperiment)
data(subR25pept)
# Simulate missing value imputation
SummarizedExperiment::assay(subR25pept[[1]])[which(is.na(assay(subR25pept[[1]])))] <- 0

qData <- as.matrix(SummarizedExperiment::assay(subR25pept[[1]]))
sTab <- SummarizedExperiment::colData(subR25pept)
limma <- limmaCompleteTest(qData, sTab)
pal <- ExtendPalette(2, "Dark2")
hc_logFC_DensityPlot(limma$logFC, th_logFC = 1, pal = pal)

}
\author{
Samuel Wieczorek
}
