% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapForMissingValues}
\alias{heatmapForMissingValues}
\title{xxx}
\usage{
heatmapForMissingValues(
  x,
  col = NULL,
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{x}{xxx}

\item{col}{xxx}

\item{srtCol}{xxx}

\item{labCol}{xxx}

\item{labRow}{xxx}

\item{key}{xxx}

\item{key.title}{xxx}

\item{main}{xxx}

\item{ylab}{xxx}
}
\value{
A heatmap
}
\description{
xxx
}
