% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{inner.sum}
\alias{inner.sum}
\title{Sum-based aggregation}
\usage{
inner.sum(pepData, X)
}
\arguments{
\item{pepData}{A \code{matrix} containing the peptide intensities.}

\item{X}{A \code{matrix} acting as an adjacency matrix.}
}
\value{
A \code{matrix} containing the aggregated values.
}
\description{
Aggregation using sum method.
}
\examples{
\dontrun{
library(QFeatures)
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept[seq_len(10)]
X <- BuildAdjacencyMatrix(obj[[length(obj)]])
i.sum <- inner.sum(assay(obj[[length(obj)]]), X)
}

}
\author{
Samuel Wieczorek
}
