% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Metacell_DIA_NN}
\alias{Metacell_DIA_NN}
\title{Sets the metacell dataframe for datasets which are from Dia-NN software}
\usage{
Metacell_DIA_NN(qdata, conds, df, level = NULL)
}
\arguments{
\item{qdata}{An object of class \code{MsnSet}}

\item{conds}{xxx}

\item{df}{A list of integer xxxxxxx}

\item{level}{xxx}
}
\value{
xxxxx
}
\description{
Actually, this function uses the generic function to generate metacell info
}
\examples{
data(subR25pept)
conds <- design.qf(subR25pept)$Condition
qdata <- SummarizedExperiment::assay(subR25pept[[2]])
df <- Metacell_DIA_NN(qdata, conds, df, level = "peptide")


}
\author{
Samuel Wieczorek
}
