% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Metacell_maxquant}
\alias{Metacell_maxquant}
\title{Sets the metacell dataframe}
\usage{
Metacell_maxquant(qdata, conds, df = NULL, level = NULL)
}
\arguments{
\item{qdata}{An object of class \code{MsnSet}}

\item{conds}{A 1-col dataframe with the condition associated to each sample.}

\item{df}{A dataframe with the same dimension as qdata containing the metacell.}

\item{level}{A string designing the type of entity/pipeline.
Available values are: \code{peptide}, \code{protein}}
}
\value{
NA
}
\description{
Initial conversion rules for maxquant
|------------|-----------------------|--------|
| Quanti     |     Identification    |    Tag |
|------------|-----------------------|--------|
|  == 0      |       whatever        |    2.0 |
|  > 0       |       'By MS/MS'      |    1.1 |
|  > 0       |      'By matching'    |    1.2 |
|  > 0       |       unknown col     |    1.0 |
|------------|-----------------------|--------|
}
\examples{
data(subR25pept)
conds <- design.qf(subR25pept)$Condition
qdata <- SummarizedExperiment::assay(subR25pept[[1]])
df2 <- Metacell_maxquant(qdata, conds, level = "peptide")

}
\author{
Samuel Wieczorek
}
