% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Metacell_proline}
\alias{Metacell_proline}
\title{Sets the metacell dataframe for datasets which are from Proline software}
\usage{
Metacell_proline(qdata, conds, df = NULL, level = NULL)
}
\arguments{
\item{qdata}{An object of class \code{MsnSet}}

\item{conds}{A 1-col dataframe with the condition associated to each sample.}

\item{df}{A dataframe with the same dimension as qdata containing the metacell.}

\item{level}{A string designing the type of entity/pipeline.
Available values are: \code{peptide}, \code{protein}}
}
\value{
NA
}
\description{
In the quantitative columns, a missing value is identified by no value rather
than a value equal to 0.

In these datasets, the metacell info is computed from the 'PSM count' columns.

Conversion rules
Initial conversion rules for proline
|--------------|-----------------|-----|
| Quanti       |    PSM count    | Tag |
|--------------|-----------------|-----|
|  == 0 | N.A. |   whatever      | 2.0 |
|  > 0         |    > 0          | 1.1 |
|  > 0         |    == 0         | 1.2 |
|  > 0         |   unknown col   | 1.0 |
|--------------|-----------------|-----|
}
\examples{
data(subR25pept)
conds <- design.qf(subR25pept)$Condition
qdata <- SummarizedExperiment::assay(subR25pept[[1]])
df <- Metacell_proline(qdata, conds, level = "peptide")


}
\author{
Samuel Wieczorek
}
