% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_QFeatures_Addons.R
\name{NAIsZero}
\alias{NAIsZero}
\title{xxx}
\usage{
NAIsZero(obj, i)
}
\arguments{
\item{obj}{An instance of QFeatures class}

\item{i}{xxx}
}
\value{
An instance of QFeatures class
}
\description{
xxx
}
