% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_Params_Addons.R, R/metacell.R,
%   R/DaparToolshed_QFeatures_Addons.R
\name{paramshistory,QFeatures-method}
\alias{paramshistory,QFeatures-method}
\alias{paramshistory,SummarizedExperiment-method}
\alias{paramshistory<-}
\alias{GetMetacellTags,QFeatures-method}
\alias{GetMetacellTags,SummarizedExperiment-method}
\alias{QFeatures-accessors}
\alias{qMetacell,QFeatures-method}
\alias{qMetacell<-}
\alias{GetUniqueTags,QFeatures-method}
\alias{GetUniqueTags,SummarizedExperiment-method}
\alias{.GetMetadataSlot}
\alias{.GetRowdataSlot}
\alias{ConnectedComp,QFeatures-method}
\alias{ConnectedComp,SummarizedExperiment-method}
\alias{ConnectedComp<-}
\alias{typeDataset,QFeatures-method}
\alias{typeDataset,SummarizedExperiment-method}
\alias{typeDataset<-}
\alias{idcol,QFeatures-method}
\alias{idcol,SummarizedExperiment-method}
\alias{idcol<-}
\alias{parentProtId,QFeatures-method}
\alias{parentProtId,SummarizedExperiment-method}
\alias{parentProtId<-}
\alias{analysis,QFeatures-method}
\alias{analysis,SummarizedExperiment-method}
\alias{analysis<-}
\alias{version,QFeatures-method}
\alias{version<-}
\alias{design.qf,QFeatures-method}
\alias{design.qf<-}
\alias{mainAssay}
\alias{HypothesisTest,QFeatures-method}
\alias{HypothesisTest,SummarizedExperiment-method}
\alias{HypothesisTest<-}
\alias{DifferentialAnalysis,QFeatures-method}
\alias{DifferentialAnalysis,SummarizedExperiment-method}
\alias{DifferentialAnalysis<-}
\alias{names_metacell,QFeatures-method}
\alias{names_metacell,SummarizedExperiment-method}
\alias{names_metacell<-}
\title{List of metacell tags}
\usage{
\S4method{paramshistory}{QFeatures}(object, i, slotName = "paramshistory")

\S4method{paramshistory}{SummarizedExperiment}(object, slotName = "paramshistory")

paramshistory(object, i, slotName = "paramshistory") <- value

\S4method{GetMetacellTags}{QFeatures}(object, i, ...)

\S4method{GetMetacellTags}{SummarizedExperiment}(object, ...)

\S4method{qMetacell}{QFeatures}(object, i)

\S4method{qMetacell}{SummarizedExperiment}(object)

qMetacell(object, i, slotName = "qMetacell") <- value

\S4method{GetUniqueTags}{QFeatures}(object, i)

\S4method{GetUniqueTags}{SummarizedExperiment}(object)

.GetMetadataSlot(object, slotName = NULL)

.GetRowdataSlot(object, slotName = NULL)

\S4method{ConnectedComp}{QFeatures}(object, i, slotName = "ConnectedComp")

\S4method{ConnectedComp}{SummarizedExperiment}(object, slotName = "ConnectedComp")

ConnectedComp(object, i, slotName = "ConnectedComp") <- value

\S4method{typeDataset}{QFeatures}(object, i, slotName = "typeDataset")

\S4method{typeDataset}{SummarizedExperiment}(object, slotName = "typeDataset")

typeDataset(object, i, slotName = "typeDataset") <- value

\S4method{idcol}{QFeatures}(object, i, slotName = "idcol")

\S4method{idcol}{SummarizedExperiment}(object, slotName = "idcol")

idcol(object, i, slotName = "idcol") <- value

\S4method{parentProtId}{QFeatures}(object, i, slotName = "parentProtId")

\S4method{parentProtId}{SummarizedExperiment}(object, slotName = "parentProtId")

parentProtId(object, i, slotName = "parentProtId") <- value

\S4method{analysis}{QFeatures}(object, i, slotName = "analysis")

\S4method{analysis}{SummarizedExperiment}(object, slotName = "analysis")

analysis(object, i, slotName = "analysis") <- value

\S4method{version}{QFeatures}(object, slotName = "version")

version(object, slotName = "version") <- value

\S4method{design.qf}{QFeatures}(object, slotName = "design")

design.qf(object, slotName = "design") <- value

mainAssay(object)

\S4method{HypothesisTest}{QFeatures}(object, i, slotName = "HypothesisTest")

\S4method{HypothesisTest}{SummarizedExperiment}(object, slotName = "HypothesisTest")

HypothesisTest(object, i) <- value

\S4method{DifferentialAnalysis}{QFeatures}(object, i, slotName = "DifferentialAnalysis")

\S4method{DifferentialAnalysis}{SummarizedExperiment}(object, slotName = "DifferentialAnalysis")

DifferentialAnalysis(object, i) <- value

\S4method{names_metacell}{QFeatures}(object, i, slotName = "names_metacell")

\S4method{names_metacell}{SummarizedExperiment}(object, slotName = "names_metacell")

names_metacell(object, i, slotName = "names_metacell") <- value
}
\arguments{
\item{object}{An instance of class \code{SummarizedExperiment} or \code{QFeatures}.}

\item{i}{The index or name of the assays to extract the quantitative
metadata from. All must have a rowdata variable named as \code{slotName}}

\item{slotName}{xxx}

\item{value}{xxx}

\item{...}{xxx}
}
\value{
A vector of tags..

NA

NA

NA

NA

NA

NA

NA

NA

NA

NA
}
\description{
This function gives the list of metacell tags available.
\itemize{
\item onlyPresent: In this case, the function gives the tags found in a dataset.
In addition, and w.r.t to the hierarchy of tags, if all leaves of a node are
present, then the tag corresponding to this node is added.
}

These names are common to all assays contained in the object. This is why
they are stored in the global metadata. This function is used whenever it i
s necessary to (re)detect MEC and POV (new dataset or when post processing
protein qMetacell after aggregation)
}
\details{
Additional slots for rowdata of a \code{SummarizedExperiment} object:
\itemize{
\item qMetacell: xxx
}

Additional slots for Metadata for a \code{QFeatures} object:
\itemize{
\item xxx: xxxx
}

Additional slots for Metadata for a \code{SummarizedExperiment} object:
\itemize{
\item qMetacell: xxxx
\item parentProtId: xxx
\item idcol: xxxx
\item typeDataset: xxx
}
}
\section{Quantitative metadata}{


Default slotName is \code{"qMetacell"}.
The value is an adjacency matrix with row and column names. The
matrix will be coerced to compressed, column-oriented sparse
matrix (class \code{dgCMatrix}) as defined in the \code{Matrix} package,
as generaled by the \code{\link[=sparseMatrix]{sparseMatrix()}} constructor.
}

\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept
GetMetacellTags(obj, 1, level="peptide")
GetMetacellTags(obj, 1, level="peptide", onlyPresent=TRUE)

data(ft, package='DaparToolshed')
design.qf(ft)
}
\author{
Samuel Wieczorek
}
