% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filter_qMetadata.R
\name{qMetacell-filter}
\alias{qMetacell-filter}
\alias{qMetacellFilteringScope}
\alias{SymFilteringOperators}
\alias{qMetacellWholeMatrix}
\alias{qMetacellWholeLine}
\alias{qMetacellOnConditions}
\title{Search lines which respects request on one or more conditions.}
\usage{
qMetacellFilteringScope()

SymFilteringOperators()

qMetacellWholeMatrix(
  object,
  cmd,
  pattern,
  percent = "Percentage",
  th,
  operator
)

qMetacellWholeLine(object, cmd, pattern)

qMetacellOnConditions(
  object,
  cmd,
  mode,
  pattern,
  conds,
  percent = "Percentage",
  operator,
  th
)
}
\arguments{
\item{object}{xxx}

\item{cmd}{Available values are:}

\item{pattern}{ss}

\item{percent}{xxx}

\item{th}{The threshold to apply}

\item{operator}{String for operator to use. List of operators is available
with \code{SymFilteringOperators()}.}

\item{mode}{xxx}

\item{conds}{xxx}
}
\value{
NA

A vector of filtering scopes

A vector of operators

NA

NA

NA
}
\description{
This function looks for the lines that respect the request in either all
conditions or at least one condition.
}
\examples{
data(ft, package='DaparToolshed')
obj <- ft[[1]]
level <- typeDataset(obj)
pattern <- "Missing"
mask <- match.metacell(
    metadata = qMetacell(obj),
    pattern = pattern,
    level = level
)
percent <- FALSE
th <- 3
op <- ">="
cmd <- 'delete' 
ind <- qMetacellWholeMatrix(obj, cmd, pattern, percent, th, op)

data(ft, package='DaparToolshed')
ind <- qMetacellWholeLine(obj, cmd, pattern)

conds <- design.qf(ft)$Condition
op <- ">="
th <- 0.5
percent <- "Percentage"
mode <- "AllCond"
ind <- qMetacellOnConditions(obj, cmd, mode, pattern, conds, percent, op, th)

qMetacellFilteringScope()

SymFilteringOperators()

data(ft_na)
obj <- ft_na[[1]]

}
