% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_FunctionFiltering.R,
%   R/DaparToolshed_filter_qMetadata.R
\name{SymFilteringOperators}
\alias{SymFilteringOperators}
\alias{qMetacell-filter}
\alias{qMetacellFilteringScope}
\alias{qMetacellWholeMatrix}
\alias{qMetacellWholeLine}
\alias{qMetacellOnConditions}
\title{Search lines which respects request on one or more conditions.}
\usage{
SymFilteringOperators()

qMetacellFilteringScope()

qMetacellWholeMatrix(
  object,
  cmd,
  pattern,
  percent = "Percentage",
  th,
  operator
)

qMetacellWholeLine(object, cmd, pattern)

qMetacellOnConditions(
  object,
  cmd,
  mode,
  pattern,
  conds,
  percent = "Percentage",
  operator,
  th
)
}
\arguments{
\item{object}{An instance of the class \code{SummarizedExperiment}}

\item{cmd}{A \code{character(1)} indicating the action to perform.
Either "keep" or "delete".}

\item{pattern}{A \code{character()} indicating the tag pattern of interest.}

\item{percent}{A \code{character()} indicating whether the threshold represent an
absolute value ("Count") or a percentage ("Percentage").}

\item{th}{The threshold to apply}

\item{operator}{String for operator to use. List of operators is available
with 'SymFilteringOperators()'.}

\item{mode}{A \code{character(1)} indicating how the task is performed.
Either "AllCond" or "AtLeastOneCond".}

\item{conds}{A vector of conditions in the dataset.}
}
\value{
A vector of operators

NA

A vector of filtering scopes

NA

NA

NA
}
\description{
This function looks for the lines that respect the request in either all
conditions or at least one condition.
}
\examples{
SymFilteringOperators()

data(subR25prot)
obj <- subR25prot[[1]]
level <- typeDataset(obj)
pattern <- "Missing"
mask <- match.metacell(
    metadata = qMetacell(obj),
    pattern = pattern,
    level = level
)
percent <- FALSE
th <- 3
op <- ">="
cmd <- 'delete' 
ind <- qMetacellWholeMatrix(obj, cmd, pattern, percent, th, op)

data(subR25prot)
ind <- qMetacellWholeLine(obj, cmd, pattern)

conds <- design.qf(subR25prot)$Condition
op <- ">="
th <- 0.5
percent <- "Percentage"
mode <- "AllCond"
ind <- qMetacellOnConditions(obj, cmd, mode, pattern, conds, percent, op, th)

qMetacellFilteringScope()

data(subR25prot)
obj <- subR25prot[[1]]

}
