% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_aggregate.R
\name{RunAggregation}
\alias{RunAggregation}
\title{xxx}
\usage{
RunAggregation(
  qf = NULL,
  includeSharedPeptides = "Yes_As_Specific",
  operator = NULL,
  considerPeptides = "allPeptides",
  topN = NULL
)
}
\arguments{
\item{includeSharedPeptides}{A boolean}

\item{operator}{xxx}

\item{considerPeptides}{xxx}

\item{topN}{In case of "top n peptides', specify then number of peptides}
}
\description{
xxx
A short description...
}
\examples{
data(ft, package='DaparToolshed')
ft
RunAggregation(ft, "Yes_As_Specific", 'colSumsMat', 'allPeptides')
RunAggregation(ft, "Yes_As_Specific", operator, 'allPeptides')
RunAggregation(ft, "No", 'colSumsMat', 'allPeptides')
RunAggregation(ft, "No", operator, 'allPeptides')

}
