% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SetHistory}
\alias{SetHistory}
\title{Standardize names}
\usage{
SetHistory(obj.se, history)
}
\arguments{
\item{obj.se}{An instance of the class \code{SummarizedExperiment}}

\item{history}{A \code{data.frame()}}
}
\description{
Standardize names
}
\examples{
data(subR25prot)
history <- GetHistory(subR25prot[[1]])
history[3, ] <- 
SetHistory(subR25prot[[1]], history)

}
\author{
Samuel Wieczorek
}
