% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatrix.R
\name{splitAdjacencyMat}
\alias{splitAdjacencyMat}
\title{splits an adjacency matrix into specific and shared}
\usage{
splitAdjacencyMat(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of two adjacency matrices
}
\description{
Method to split an adjacency matrix into specific and shared
}
\examples{
data(subR25pept)
X <- BuildAdjacencyMatrix(subR25pept[[2]])
ll <- splitAdjacencyMat(X)

}
\author{
Samuel Wieczorek
}
