% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R, R/design_utils.R
\name{test.design}
\alias{test.design}
\title{Check if xxxxxx}
\usage{
test.design(tab)

test.design(tab)
}
\arguments{
\item{tab}{A data.frame which correspond to xxxxxx}
}
\value{
A list of two items

A list of two items
}
\description{
Check if xxxxxx

Check if xxxxxx
}
\examples{
data(Exp1_R25_pept, package='DaparToolshedData')
test.design(colData(Exp1_R25_pept)[, seq(3)])

data(Exp1_R25_pept, package="DaparToolshedData")
test.design(design.qf(Exp1_R25_pept)[, seq_len(3)])

}
\author{
Thomas Burger, Samuel Wieczorek
}
